% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic.R
\name{Abiotic.forced}
\alias{Abiotic.forced}
\title{Set up exogenous variables for abiotic forcing}
\usage{
\method{Abiotic}{forced}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
Implements \link{Abiotic} for abiotic forcing
}
