% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic.R
\name{Abiotic}
\alias{Abiotic}
\title{Set up exogenous variables for abiotic forcing}
\usage{
Abiotic(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$ABIOTIC}.
}
