% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{approx_equal}
\alias{approx_equal}
\title{Check if two numeric values are approximately equal}
\usage{
approx_equal(a, b, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{a}{a \link{numeric} object}

\item{b}{a \link{numeric} object}

\item{tol}{the numeric tolerance}
}
\value{
a logical value
}
\description{
Check if two numeric values are approximately equal
}
