% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{AreaSprayEffects.null}
\alias{AreaSprayEffects.null}
\title{Set up mass spraying}
\usage{
\method{AreaSprayEffects}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{AreaSprayEffects} for the null model (do nothing)
}
