% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{AreaSprayEffectSizes}
\alias{AreaSprayEffectSizes}
\title{Methods for mass spraying}
\usage{
AreaSprayEffectSizes(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$AREA_SPRAY}.
}
