% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsb-null.R
\name{ATSB.null}
\alias{ATSB.null}
\title{Modify baseline values due to ATSB}
\usage{
\method{ATSB}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a \link{list}
}
\description{
Implements \link{ATSB} for the null model of ATSB (do nothing)
}
