% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atsb-interface.R
\name{ATSB}
\alias{ATSB}
\title{Modify baseline bionomic values due to ATSB}
\usage{
ATSB(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$ATSBpar}.
}
