% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-basic.R
\name{attach_inits_H.basic}
\alias{attach_inits_H.basic}
\title{Attach initial values to parameter list}
\usage{
\method{attach_inits_H}{basic}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{attach_inits_H} for the basic model.
}
