% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{attach_inits_H}
\alias{attach_inits_H}
\title{Attach the initial variable values to the parameter list}
\usage{
attach_inits_H(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Hpar}.
}
