% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_blood.R
\name{AvailableBlood.forced}
\alias{AvailableBlood.forced}
\title{Compute availability of local humans for blood feeding}
\usage{
\method{AvailableBlood}{forced}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
