% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_blood.R
\name{AvailableBlood}
\alias{AvailableBlood}
\title{Update the availability of blood hosts}
\usage{
AvailableBlood(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$BFpar}.
}
