% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_habitat.R
\name{AvailableHabitat.forced}
\alias{AvailableHabitat.forced}
\title{Compute total availability of aquatic habitats}
\usage{
\method{AvailableHabitat}{forced}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of aquatic habitats
}
