% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_habitat.R
\name{AvailableHabitat}
\alias{AvailableHabitat}
\title{Update the availability of aquatic habitats}
\usage{
AvailableHabitat(pars)
}
\arguments{
\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$EGGpar}.
}
