% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_habitat.R
\name{AvailableHabitat.static}
\alias{AvailableHabitat.static}
\title{Compute total availability of aquatic habitats}
\usage{
\method{AvailableHabitat}{static}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of aquatic habitats for the static model (do nothing)
}
