% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_sugar.R
\name{AvailableSugar.forced}
\alias{AvailableSugar.forced}
\title{Compute total availability of sugar}
\usage{
\method{AvailableSugar}{forced}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of sugar
}
