% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-lemenach.R
\name{BedNetEffects.lemenach}
\alias{BedNetEffects.lemenach}
\title{Modify baseline values due to vector control}
\usage{
\method{BedNetEffects}{lemenach}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{BedNetEffectSizes} for the Le Menach ITN model of vector control
}
