% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{BedNetEffects.null}
\alias{BedNetEffects.null}
\title{Bed net ownership}
\usage{
\method{BedNetEffects}{null}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Implements \link{BedNetEffects} for the null model of bed nets (do nothing)
}
