% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNetEffects}
\alias{BedNetEffects}
\title{Modify variables or parameters, called from VectorControlEffects}
\usage{
BedNetEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$ITNeff}.
}
