% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-lemenach.R
\name{BedNetEffectSizes.lemenach}
\alias{BedNetEffectSizes.lemenach}
\title{Modify baseline values due to vector control}
\usage{
\method{BedNetEffectSizes}{lemenach}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{BedNetEffectSizes} for the Le Menach ITN model of vector control
}
