% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{BedNetEffectSizes.null}
\alias{BedNetEffectSizes.null}
\title{Bed net ownership}
\usage{
\method{BedNetEffectSizes}{null}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{BedNetEffectSizes} for the null model of bed nets (do nothing)
}
