% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNetEffectSizes}
\alias{BedNetEffectSizes}
\title{Modify baseline bionomic parameters, called from VectorControlEffectSizes}
\usage{
BedNetEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$ITNefsz}.
}
