% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-lemenach.R
\name{BedNets.lemenach}
\alias{BedNets.lemenach}
\title{Modify baseline values due to vector control}
\usage{
\method{BedNets}{lemenach}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a named \link{list}
}
\description{
Implements \link{BedNets} for the Le Menach ITN model of vector control
}
