% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{BedNets.null}
\alias{BedNets.null}
\title{Modify baseline values due to bed nets}
\usage{
\method{BedNets}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{BedNets} for the null model of bed nets (do nothing)
}
