% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{BedNets}
\alias{BedNets}
\title{Modify baseline bionomic values due to bed nets}
\usage{
BedNets(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$ITNpar}.
}
