% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{Behavior.null}
\alias{Behavior.null}
\title{Methods for dynamic human and mosquito behaviors}
\usage{
\method{Behavior}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Behavior} for the null model (no changes)
}
