% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{Behavior}
\alias{Behavior}
\title{Methods for dynamic human and mosquito behaviors}
\usage{
Behavior(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$BEHAVIOR}.
}
