% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{Bionomics}
\alias{Bionomics}
\title{Set bionomic parameter rates relative to baseline}
\usage{
Bionomics(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This calls Mbionomics and Lbionmics for each species. This function
resets bionomic parameters to their pre-control baseline value, which can later be
modified by vector control. In some models, the pre-control baseline is computed in
here as a function of resource availability.
}
