% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{Births.numeric}
\alias{Births.numeric}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{Births}{numeric}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{Births} when \code{y} is numeric
}
