% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-births.R
\name{Births.zero}
\alias{Births.zero}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{Births}{zero}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of zeros
}
\description{
Implements \link{Births} for static population models
}
