% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_seeking.R
\name{CareSeeking.null}
\alias{CareSeeking.null}
\title{Set the values of exogenous variables describing care seeking}
\usage{
\method{CareSeeking}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{CareSeeking} for the null model of care seeking (do nothing)
}
