% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_seeking.R
\name{CareSeeking}
\alias{CareSeeking}
\title{Set the values of exogenous variables describing care seeking}
\usage{
CareSeeking(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$CARE_SEEKING}.
}
