% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkIt}
\alias{checkIt}
\title{Check the length of an input value}
\usage{
checkIt(x, lng, type = "numeric", fixit = TRUE)
}
\arguments{
\item{x}{a \link{numeric} object}

\item{lng}{a \link{numeric} object}

\item{type}{a \link{character} string specifying required typeof}

\item{fixit}{a \link{logical} value, if TRUE force length to lng}
}
\value{
a \link{logical} value
}
\description{
Check the length of an input value
}
