% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil.R
\name{Civil.forced}
\alias{Civil.forced}
\title{Set up exogenous variables for civil forcing}
\usage{
\method{Civil}{forced}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
Implements \link{Civil} for civil forcing
}
