% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil.R
\name{Civil}
\alias{Civil}
\title{Set up exogenous variables for civil forcing}
\usage{
Civil(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$CIVIL}.
}
