% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{Clinic.null}
\alias{Clinic.null}
\title{Methods for distributing interventions during clinical visits}
\usage{
\method{Clinic}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Clinic} for the null model (do nothing)
}
