% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_habitat.R
\name{compute_AvailableHabitat}
\alias{compute_AvailableHabitat}
\title{Compute total availability of aquatic habitats}
\usage{
compute_AvailableHabitat(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of aquatic habitats
}
