% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_blood.R
\name{compute_AvailableHosts}
\alias{compute_AvailableHosts}
\title{Compute availability blood hosts of the i^th species}
\usage{
compute_AvailableHosts(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
