% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{compute_B}
\alias{compute_B}
\title{Compute the total availability of blood hosts}
\usage{
compute_B(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
