% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_beta}
\alias{compute_beta}
\title{Compute beta, the biting distribution matrix}
\usage{
compute_beta(H, wts_f, TaR)
}
\arguments{
\item{H}{the human population size}

\item{wts_f}{the blood feeding search weights}

\item{TaR}{(time at risk), a \link{matrix}  dimensions \code{nPatches} by \code{nStrata}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Compute beta, the biting distribution matrix
}
