% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{compute_calU}
\alias{compute_calU}
\title{Compute calU}
\usage{
compute_calU(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object}
}
\value{
a \link{list}
}
\description{
Computes the egg laying matrices
}
