% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_EIR.aqua}
\alias{compute_EIR.aqua}
\title{Compute the EIR}
\usage{
\method{compute_EIR}{aqua}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{matrix}
}
\description{
Implements \link{compute_EIR} for aquatic ODEs
}
