% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_EIR.mosy_dde}
\alias{compute_EIR.mosy_dde}
\title{Compute the EIR}
\usage{
\method{compute_EIR}{mosy_dde}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{vector}
}
\description{
Implements \link{compute_EIR} for mosquito DDEs
}
