% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_EIR}
\alias{compute_EIR}
\title{Compute EIR}
\usage{
compute_EIR(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars, }{a \link{list}}
}
\value{
\link{list}
}
\description{
This function computes the EIR for each stratum of each host species
}
