% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_EIR_i}
\alias{compute_EIR_i}
\title{Compute the EIR for the ith point in time}
\usage{
compute_EIR_i(i, deout, pars)
}
\arguments{
\item{i}{an \link{integer}}

\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Using the output of deSolve,
compute the EIR at one point in time
}
