% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_EIR_ix}
\alias{compute_EIR_ix}
\title{Compute the EIR}
\usage{
compute_EIR_ix(ix, deout, pars, s, i)
}
\arguments{
\item{ix}{an \link{integer}}

\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
\link{numeric}
}
\description{
Using the output of deSolve,
compute the EIR at the i^th point in time
}
