% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_fqM}
\alias{compute_fqM}
\title{Compute the fqM}
\usage{
compute_fqM(deout, pars, s)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
\link{vector}
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the fqM at every point in time
}
