% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_fqZ.mosy}
\alias{compute_fqZ.mosy}
\title{Compute the fqZ}
\usage{
\method{compute_fqZ}{mosy}(deout, pars)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{vector}
}
\description{
Implements \link{compute_fqZ} for class(pars$xde)=mosy
}
