% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_terms.R
\name{compute_fqZ}
\alias{compute_fqZ}
\title{Compute the fqZ}
\usage{
compute_fqZ(deout, pars, s)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
\link{vector}
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the fqZ at every point in time
}
