% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_fqZ_i}
\alias{compute_fqZ_i}
\title{Compute the fqZ for the ith}
\usage{
compute_fqZ_i(i, deout, pars)
}
\arguments{
\item{i}{an \link{integer}}

\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the fqZ at one point in time
}
