% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_terms.R
\name{compute_fqZ_ix}
\alias{compute_fqZ_ix}
\title{Compute the fqZ for the ith}
\usage{
compute_fqZ_ix(ix, deout, pars, s)
}
\arguments{
\item{ix}{an \link{integer}}

\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
\link{numeric}
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the fqZ at one point in time
}
