% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_kappa.human}
\alias{compute_kappa.human}
\title{Compute the kappa}
\usage{
\method{compute_kappa}{human}(deout, pars)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{vector}
}
\description{
Implements \link{compute_kappa} for class(pars$xde)=human
}
