% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_kappa.mosy_dde}
\alias{compute_kappa.mosy_dde}
\title{Compute the kappa}
\usage{
\method{compute_kappa}{mosy_dde}(deout, pars)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{vector}
}
\description{
Implements \link{compute_kappa} for class(pars$xde)=mosy_dde
}
