% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_kappa}
\alias{compute_kappa}
\title{Compute kappa}
\usage{
compute_kappa(deout, pars, i)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
\link{numeric} containing the kappa
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the kappa for each stratum
}
