% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_kappa_default}
\alias{compute_kappa_default}
\title{Compute the kappa}
\usage{
compute_kappa_default(deout, pars)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{numeric} containing the kappa
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the kappa for each stratum
}
