% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_kappa_ix}
\alias{compute_kappa_ix}
\title{Compute kappa once}
\usage{
compute_kappa_ix(ix, deout, pars, i)
}
\arguments{
\item{ix}{an \link{integer}}

\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list} containing the kappa
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the kappa at a point in time
}
