% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_kappa_ty}
\alias{compute_kappa_ty}
\title{Compute kappa, ty}
\usage{
compute_kappa_ty(t, y, pars, i)
}
\arguments{
\item{t}{\link{numeric}}

\item{y}{\link{vector} the state variables}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list} containing the kappa
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the kappa at a point in time
}
