% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{compute_local_frac}
\alias{compute_local_frac}
\title{Compute the fraction of bites occurring on local hosts}
\usage{
compute_local_frac(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the local_frac
}
