% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_NI.cohort}
\alias{compute_NI.cohort}
\title{Compute the NI}
\usage{
\method{compute_NI}{cohort}(deout, pars)
}
\arguments{
\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{vector}
}
\description{
Implements \link{compute_NI} for class(pars$xde)=cohort
}
