% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_NI_i}
\alias{compute_NI_i}
\title{Compute NI once}
\usage{
compute_NI_i(i, deout, pars)
}
\arguments{
\item{i}{an \link{integer}}

\item{deout}{a matrix, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
a \link{list} containing the NI
}
\description{
Using the output of \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede},
compute the NI at a point in time
}
