% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{compute_Q}
\alias{compute_Q}
\title{Compute total availability of aquatic habitats}
\usage{
compute_Q(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of aquatic habitats
}
