% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_terms.human}
\alias{compute_terms.human}
\title{Compute dynamical terms}
\usage{
\method{compute_terms}{human}(varslist, deout, pars, s, i)
}
\arguments{
\item{varslist}{a \link{list} with variable names attached}

\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
\link{list}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{xde_solve.human}
}
