% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_terms.na}
\alias{compute_terms.na}
\title{Compute dynamical terms}
\usage{
\method{compute_terms}{na}(varslist, deout, pars)
}
\arguments{
\item{varslist}{a \link{list} with variable names attached}

\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{xde_solve} that
don't have any relevant dynamical terms
}
