% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_terms}
\alias{compute_terms}
\title{Compute dynamical terms}
\usage{
compute_terms(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve compute the dynamical terms every point in time. This method dispatches on the type of \code{pars$compute}.
}
