% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_terms.xde}
\alias{compute_terms.xde}
\title{Compute dynamical terms}
\usage{
\method{compute_terms}{xde}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{xde_solve.ode} or \code{xde_solve.dde}
}
