% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_terms.R
\name{compute_terms_steady}
\alias{compute_terms_steady}
\title{Compute dynamical terms}
\usage{
compute_terms_steady(varslist, y_eq, pars, s, i)
}
\arguments{
\item{varslist}{a \link{list} with variable names attached}

\item{y_eq}{a \link{matrix}, the output of \code{rootSolve}}

\item{pars}{a \link{list}}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{root_solve}
}
