% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{compute_vars_aqua}
\alias{compute_vars_aqua}
\title{Differential equation models for aquatic mosquito populations}
\usage{
compute_vars_aqua(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
\strong{pars} a \link{list}
}
\description{
Compute everything but the derivatives for an aquatic mosquito-only
differential equation model
}
