% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_variables.R
\name{compute_vars_mosy}
\alias{compute_vars_mosy}
\title{Compute other variables at time t}
\usage{
compute_vars_mosy(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}
the appropriate adult mosquito model}
}
\value{
\strong{pars} a \link{list}
}
\description{
Compute everything but the derivatives for a mosquito-only
differential equation model
}
