% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{compute_W}
\alias{compute_W}
\title{Compute host availability}
\usage{
compute_W(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
