% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{computeB.dynamic}
\alias{computeB.dynamic}
\title{Static model for hosts}
\usage{
\method{computeB}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{computeB} for a dynamic model
}
