% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{computeB}
\alias{computeB}
\title{Set the availability of hosts}
\usage{
computeB(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$Bpar}. It should
compute the availability of all blood hosts at time \code{t}
}
