% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{computeQ.dynamic}
\alias{computeQ.dynamic}
\title{Static model for hosts}
\usage{
\method{computeQ}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{computeQ} for a static model
}
