% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{computeQ}
\alias{computeQ}
\title{Set the availability of aquatic habitats}
\usage{
computeQ(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$Qpar}. It should
compute the availability of aquatic habitats at time \code{t}
}
