% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{computeQ.static}
\alias{computeQ.static}
\title{Static model for hosts}
\usage{
\method{computeQ}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{computeQ} for a static model
}
