% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{computeW}
\alias{computeW}
\title{Set the availability of hosts}
\usage{
computeW(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$Wpar}. It should
compute the availability of the pathogen's hosts for blood feeding at time \code{t}
}
