% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{computeW.static}
\alias{computeW.static}
\title{Static model for hosts}
\usage{
\method{computeW}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{computeW} for a static model
}
