% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{Control.forced}
\alias{Control.forced}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{Control}{forced}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
Implements \link{Control} for a model with some control
}
