% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{Control}
\alias{Control}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
Control(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$CONTROL}.
}
