% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dde2ode_MYZ.dde}
\alias{dde2ode_MYZ.dde}
\title{Convert a model from dde to the corresponding ode}
\usage{
\method{dde2ode_MYZ}{dde}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
If it is a dde, return the corresponding ode
}
