% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dde2ode_MYZ}
\alias{dde2ode_MYZ}
\title{Convert a model from dde to the corresponding ode}
\usage{
dde2ode_MYZ(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar$xde}
}
