% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{dEIPdt}
\alias{dEIPdt}
\title{Compute the derivative of the EIP as a function of time}
\usage{
dEIPdt(t, EIPmod)
}
\arguments{
\item{t}{current simulation time}

\item{EIPmod}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
This method dispatches on the type of \code{EIPmod}.
}
