% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Development.null}
\alias{Development.null}
\title{Set up developments}
\usage{
\method{Development}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Development} for the null model (do nothing)
}
