% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Development}
\alias{Development}
\title{Set up developments}
\usage{
Development(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$DEVELOPMENT}.
}
