% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-null.R
\name{dHdt.null}
\alias{dHdt.null}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{dHdt}{null}(t, y, pars)
}
\value{
a \link{numeric} vector of length 0
}
\description{
Implements \link{dHdt} for the null model.
}
