% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{dHdt.numeric}
\alias{dHdt.numeric}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{dHdt}{numeric}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{dHdt} when \code{y} is numeric
}
