% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{dHdt.trace}
\alias{dHdt.trace}
\title{Dynamics of human population denominators}
\usage{
\method{dHdt}{trace}(t, y, dy, pars, D = NULL)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{D}{a demography matrix, which should be of dimensions \code{nStrata} by \code{nStrata}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{dHdt} for the forced (trace) model.
}
