% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_inverse}
\alias{diag_inverse}
\title{Invert a diagonal matrix}
\usage{
diag_inverse(x)
}
\arguments{
\item{x}{a \link{numeric} vector}
}
\value{
a diagonal \link{matrix}
}
\description{
Invert a diagonal matrix which is passed as a vector. If any
elements are zero, set them to one.
}
