% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{DistributeBedNets.null}
\alias{DistributeBedNets.null}
\title{Distribute bed nets}
\usage{
\method{DistributeBedNets}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{DistributeBedNets} for the null model of bed nets (do nothing)
}
