% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{DistributeBedNets}
\alias{DistributeBedNets}
\title{Distribute bed nets, called from Control(VectorControl)}
\usage{
DistributeBedNets(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$ITNdist}.
}
