% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dMYZdt}
\alias{dMYZdt}
\title{Derivatives for adult mosquitoes}
\usage{
dMYZdt(t, y, pars, Lambda, kappa, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
