% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{dMYZdt.Ztrace}
\alias{dMYZdt.Ztrace}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dMYZdt}{Ztrace}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the Ztrace (forced emergence) model.
}
