% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dXdt.SISdX}
\alias{dXdt.SISdX}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SISdX}(t, y, pars, FoI)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{FoI}{vector giving the per-capita force of infection for each strata}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SIS model, no demography.
}
