% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{dXdt.trace}
\alias{dXdt.trace}
\title{Derivatives for human population}
\usage{
\method{dXdt}{trace}(t, y, pars, FoI, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{FoI}{vector giving the per-capita force of infection for each strata}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the trace model.
}
