% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{EggLaying.simple}
\alias{EggLaying.simple}
\title{Compute eggs laid}
\usage{
\method{EggLaying}{simple}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object}
}
\value{
a \link{list}
}
\description{
Computes eggs laid for the s^th species
}
