% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{EIP}
\alias{EIP}
\title{Compute the EIP}
\usage{
EIP(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$EIPmod}.
}
