% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{EIP.static}
\alias{EIP.static}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{EIP}{static}(t, EIPmod)
}
\arguments{
\item{t}{current simulation time}

\item{EIPmod}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{EIP} for the static model (the EIP is constant)
}
