% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emerge.R
\name{Emergence}
\alias{Emergence}
\title{Compute emerging adults}
\usage{
Emergence(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object}
}
\value{
a \link{list}
}
\description{
Computes eggs laid for the s^th species
}
