% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-forced.R
\name{ExogenousForcing.forced}
\alias{ExogenousForcing.forced}
\title{Modify parameters due to exogenous forcing}
\usage{
\method{ExogenousForcing}{forced}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
Calls functions that set up exogenous variables
}
