% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogeneous-null.R
\name{ExogenousForcing.null}
\alias{ExogenousForcing.null}
\title{Modify parameters due to exogenous forcing}
\usage{
\method{ExogenousForcing}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
nothing
}
\description{
Implements \link{ExogenousForcing} for the null model of exogenous forcing (do nothing)
}
