% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-interface.R
\name{ExogenousForcing}
\alias{ExogenousForcing}
\title{Modify parameters due to exogenous forcing}
\usage{
ExogenousForcing(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$EXOpar}.
}
