% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{F_alpha}
\alias{F_alpha}
\title{Number of newly emerging adults from each larval habitat}
\usage{
F_alpha(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}
}
\value{
a [numeric] vector of length `nHabitats`
}
\description{
This method dispatches on the type of `pars$Lpar`.
}
