% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_b.hMoI}
\alias{F_b.hMoI}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_b}{hMoI}(y, pars)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_b} for the hMoI model.
}
