% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_b.SIS}
\alias{F_b.SIS}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_b}{SIS}(y, pars)
}
\arguments{
\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_b} for the SIS model.
}
