% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodfeeding.R
\name{F_beta.dynamic}
\alias{F_beta.dynamic}
\title{Biting distribution matrix}
\usage{
\method{F_beta}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Implements \link{F_beta} for static denominators
}
