% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_beta}
\alias{F_beta}
\title{Compute beta, the biting distribution matrix}
\usage{
F_beta(H, W, wts_f, TaR)
}
\arguments{
\item{H}{human / host population density}

\item{W}{human / host availability in the patches}

\item{wts_f}{the blood feeding search weights}

\item{TaR}{(time at risk), a \link{matrix}  dimensions \code{nPatches} by \code{nStrata}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Compute beta, the biting distribution matrix
}
