% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_beta.SIS}
\alias{F_beta.SIS}
\title{Biting distribution matrix}
\usage{
\method{F_beta}{SIS}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Implements \link{F_beta} for the SIS model.
}
