% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission-trace.R
\name{F_beta.trace}
\alias{F_beta.trace}
\title{Net infectiousness of human population to mosquitoes}
\usage{
\method{F_beta}{trace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list} vector of length \code{nPatches}
}
\description{
Implements \link{F_beta} for an trace model.
}
