% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_beta_lag.hMoI}
\alias{F_beta_lag.hMoI}
\title{Lagged biting distribution matrix}
\usage{
\method{F_beta_lag}{hMoI}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{is the lag value}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Implements \link{F_beta_lag} for the hybrid MoI model.
}
