% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{F_beta_lag}
\alias{F_beta_lag}
\title{Lagged biting distribution matrix}
\usage{
F_beta_lag(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{duration of lag \code{t-lag}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
