% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_beta_lag.SIP}
\alias{F_beta_lag.SIP}
\title{Lagged biting distribution matrix}
\usage{
\method{F_beta_lag}{SIP}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{duration of lag \code{t-lag}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Implements \link{F_beta_lag} for the SIP model.
}
