% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodfeeding.R
\name{F_beta_lag.static}
\alias{F_beta_lag.static}
\title{Biting distribution matrix}
\usage{
\method{F_beta_lag}{static}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Implements \link{F_beta_lag} when the availability of blood hosts is static
}
