% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{F_births.0}
\alias{F_births.0}
\title{A function that computes the human population birth rate}
\usage{
\method{F_births}{`0`}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_births} for the constant births model
}
