% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-births.R
\name{F_births.exp}
\alias{F_births.exp}
\title{Constant per-capita human population birth rate}
\usage{
\method{F_births}{exp}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_births} with a constant per-capita population birth rate
}
