% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\name{F_births.forced}
\alias{F_births.forced}
\title{Growing human population birth rate (forced)}
\usage{
\method{F_births}{forced}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_births} with a growing population birth rate
}
