% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\name{F_births.null}
\alias{F_births.null}
\title{Null human population births}
\usage{
\method{F_births}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_births} for a null model
}
