% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\name{F_births}
\alias{F_births}
\title{The human population birth rate}
\usage{
F_births(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of \code{pars$Hpar$birthF}.
}
