% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{F_eggs.GeRM}
\alias{F_eggs.GeRM}
\title{Number of eggs laid by adult mosquitoes}
\usage{
\method{F_eggs}{GeRM}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_eggs} for the generalized GeRM model.
}
