% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{F_eggs}
\alias{F_eggs}
\title{Number of eggs laid by adult mosquitoes}
\usage{
F_eggs(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
