% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{F_eip}
\alias{F_eip}
\title{Compute the eip}
\usage{
F_eip(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar$eip_par}. It should
set the values of the eip
}
