% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_EIR.hMoI}
\alias{F_EIR.hMoI}
\title{Entomological inoculation rate on human strata}
\usage{
\method{F_EIR}{hMoI}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_EIR} for the hybrid MoI model.
}
