% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission-ode.R
\name{F_EIR.ode}
\alias{F_EIR.ode}
\title{Entomological inoculation rate on human strata}
\usage{
\method{F_EIR}{ode}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_EIR} for when X is an ODE model.
}
