% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_EIR}
\alias{F_EIR}
\title{Entomological inoculation rate on human strata}
\usage{
F_EIR(t, y, pars, i, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{s}{the vector species index}
}
\value{
\link{numeric} vector of length \code{nStrata}
}
\description{
Compute the daily EIR for all the strata at time \code{t}
}
