% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodfeeding.R
\name{F_EIR_lag}
\alias{F_EIR_lag}
\title{Entomological inoculation rate on human strata at a lag}
\usage{
F_EIR_lag(t, y, pars, MosyBehavior)
}
\arguments{
\item{t}{lagged simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{MosyBehavior}{values returned by \link{MosquitoBehavior}, potentially modified by control \link{VectorControl}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Entomological inoculation rate on human strata at a lag
}
