% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{F_f}
\alias{F_f}
\title{Compute the blood feeding rate, f}
\usage{
F_f(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar$f_par}. It should
set the values of the bionomic parameters to baseline values.
}
