% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{F_foi.nb}
\alias{F_foi.nb}
\title{Null human population births}
\usage{
\method{F_foi}{nb}(eir, pars)
}
\arguments{
\item{eir}{the daily eir}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_foi} for a negative binomial model
}
