% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{F_foi}
\alias{F_foi}
\title{A model for daily FoI as a function of the daily EIR.}
\usage{
F_foi(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{the model object as a \link{list}}
}
\value{
the daily, local force of infection as a \link{numeric} vector
}
\description{
This function compute the daily local FoI as a function
of the daily EIR and effects of partial immunity.
It is computed based on a probabilistic model of exposure and
possibly including environmental heterogeneity. If a model of human / host
infection has terms that describe partial immunity, \emph{e.g.} affecting
pre-erythrocytic immunity to malaria called by \link{F_b}, those effects are implemented here.
The method dispatches on the type of \code{pars$FOIpar}.
}
