% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{F_fqM.Gtrace}
\alias{F_fqM.Gtrace}
\title{Blood feeding rate of the infective mosquito population}
\usage{
\method{F_fqM}{Gtrace}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_fqM} for the Gtrace model.
}
