% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{F_fqZ}
\alias{F_fqZ}
\title{Blood feeding rate of the infective mosquito population}
\usage{
F_fqZ(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
