% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-basic.R
\name{F_H.basic}
\alias{F_H.basic}
\title{Size of human population denominators}
\usage{
\method{F_H}{basic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the forced (basic) model.
}
