% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{F_H.dynamic}
\alias{F_H.dynamic}
\title{Size of human population denominators}
\usage{
\method{F_H}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the dynamic demography model
}
