% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_H.SIS}
\alias{F_H.SIS}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SIS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the SIS model.
}
