% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{F_H.static}
\alias{F_H.static}
\title{Size of human population denominators}
\usage{
\method{F_H}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the static model.
}
