% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{F_habitats.forced}
\alias{F_habitats.forced}
\title{A basic forcing model for habitat availability}
\usage{
\method{F_habitats}{forced}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{F_habitats} for a forced model
}
