% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{F_habitats.null}
\alias{F_habitats.null}
\title{A null model for habitat}
\usage{
\method{F_habitats}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{F_habitats} for a static model
}
