% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{F_habitats}
\alias{F_habitats}
\title{Set the values for habitat}
\usage{
F_habitats(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$QHABpar}. It should
set the values of habitat at time \code{t}
}
