% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{F_kappa.GeRM_ode}
\alias{F_kappa.GeRM_ode}
\title{Net infectiousness of human population to mosquitoes}
\usage{
\method{F_kappa}{GeRM_ode}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_kappa} for the generalized GeRM ODE model.
}
