% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission-trace.R
\name{F_kappa.trace}
\alias{F_kappa.trace}
\title{Net infectiousness of human population to mosquitoes}
\usage{
\method{F_kappa}{trace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_kappa} for when MYZ is an trace model.
}
