% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{F_nu}
\alias{F_nu}
\title{Compute the egg laying rate}
\usage{
F_nu(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar$nu_par}. It should
set the values of nu to (possibly changing) baseline value(s).
}
