% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{F_other}
\alias{F_other}
\title{Set the availability of other blood hosts}
\usage{
F_other(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$OBpar}. It should
compute the availability of other blood hosts at time \code{t}
}
