% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{F_otherblood.forced}
\alias{F_otherblood.forced}
\title{A basic forcing model for other blood hosts}
\usage{
\method{F_otherblood}{forced}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{F_otherblood} for a forced model
}
