% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{F_otherblood.null}
\alias{F_otherblood.null}
\title{Null model for other blood hosts}
\usage{
\method{F_otherblood}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{F_otherblood} for a null model
}
