% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{F_otherblood}
\alias{F_otherblood}
\title{Set the availability of other}
\usage{
F_otherblood(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$OBHpar}. It should
set the values of other availability at time \code{t}
}
