% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_pr.SIP}
\alias{F_pr.SIP}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{SIP}(varslist, pars, i)
}
\arguments{
\item{varslist}{a list with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIP model.
}
