% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{F_pr.trace}
\alias{F_pr.trace}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{trace}(varslist, pars)
}
\arguments{
\item{varslist}{a \link{list} with variables attached by name}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector numeric(0)
}
\description{
Implements \link{F_pr} for the trace model.
}
