% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{F_q.dynamic}
\alias{F_q.dynamic}
\title{Static model for human blood fraction}
\usage{
\method{F_q}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_q} for a static model
}
