% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{F_sigma.BQS}
\alias{F_sigma.BQS}
\title{Model for mosquito emigration based on resource availability}
\usage{
\method{F_sigma}{BQS}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_sigma} for a static model
}
