% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{F_sugar.forced}
\alias{F_sugar.forced}
\title{A basic forcing model for sugar}
\usage{
\method{F_sugar}{forced}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{F_sugar} for a forced model
}
