% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{F_sugar}
\alias{F_sugar}
\title{Set the availability of sugar}
\usage{
F_sugar(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$SGRpar}. It should
set the values of sugar availability at time \code{t}
}
