% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{F_sugar.static}
\alias{F_sugar.static}
\title{Basic model for sugar}
\usage{
\method{F_sugar}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_sugar} for a basic model
}
