% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{F_tau.RM}
\alias{F_tau.RM}
\title{Time spent host seeking/feeding and resting/ovipositing}
\usage{
\method{F_tau}{RM}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
\link{NULL}
}
\description{
Implements \link{F_tau} for the generalized RM model.
}
