% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_x.SIS}
\alias{F_x.SIS}
\title{Size of effective infectious human population}
\usage{
\method{F_x}{SIS}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_x] for the SIS model.
}
