% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_x_lag.SIS}
\alias{F_x_lag.SIS}
\title{Size of lagged effective infectious human population}
\usage{
\method{F_x_lag}{SIS}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{duration of lag \code{t-lag}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_x_lag} for the SIS model.
}
