% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{F_x_tau}
\alias{F_x_tau}
\title{Size of lagged effective infectious human population}
\usage{
F_x_tau(t, y, pars, tau)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}

\item{tau}{duration of lag `t-tau`}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
This method dispatches on the type of `pars$Xpar`.
}
