% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_x_tau.SIP}
\alias{F_x_tau.SIP}
\title{Size of lagged effective infectious human population}
\usage{
\method{F_x_tau}{SIP}(t, y, pars, tau)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}

\item{tau}{duration of lag `t-tau`}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_x_tau] for the SIP model.
}
