% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{F_Z.GeRM}
\alias{F_Z.GeRM}
\title{Density of infectious mosquitoes}
\usage{
\method{F_Z}{GeRM}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_Z} for the GeRM model.
}
