% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{F_Z.Gtrace}
\alias{F_Z.Gtrace}
\title{Number of infective adults in each patch}
\usage{
\method{F_Z}{Gtrace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_Z} for the Gtrace  model.
}
