% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{F_Z}
\alias{F_Z}
\title{Density of infectious mosquitoes}
\usage{
F_Z(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
