% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{F_Z.Ztrace}
\alias{F_Z.Ztrace}
\title{Number of infective adults in each patch}
\usage{
\method{F_Z}{Ztrace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_Z} for the Ztrace  model.
}
