% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{F_Z_lag}
\alias{F_Z_lag}
\title{Density of lagged infectious mosquitoes}
\usage{
F_Z_lag(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{lag}{duration of lag \code{t-lag}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
