% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{F_Z_lag.RM}
\alias{F_Z_lag.RM}
\title{Density of lagged infectious mosquitoes}
\usage{
\method{F_Z_lag}{RM}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{lag}{duration of lag \code{t-lag}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_Z_lag} for the generalized RM model.
}
