% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{F_Z_lag.Ztrace}
\alias{F_Z_lag.Ztrace}
\title{Number of infective adults in each patch}
\usage{
\method{F_Z_lag}{Ztrace}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{duration of lag \code{t-lag}}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_Z} for the Ztrace  model.
}
