% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{F_Z_tau.RM}
\alias{F_Z_tau.RM}
\title{Density of lagged infectious mosquitoes}
\usage{
\method{F_Z_tau}{RM}(t, y, pars, tau)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{tau}{duration of lag \code{t-tau}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_Z_tau} for the generalized RM model.
}
