% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{fqZ_local.basic}
\alias{fqZ_local.basic}
\title{Parasite / pathogen importation, the basic model}
\usage{
\method{fqZ_local}{basic}(fqZ, pars)
}
\arguments{
\item{fqZ}{blood meals per patch, per day}

\item{pars}{a \link{list}}
}
\value{
fqZ, a numeric \link{vector}
}
\description{
Implements \link{fqZ_local} for the basic model (do nothing)
}
