% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-null.R
\name{fqZ_local.null}
\alias{fqZ_local.null}
\title{Parasite / pathogen importation, the null model}
\usage{
\method{fqZ_local}{null}(fqZ, pars)
}
\arguments{
\item{fqZ}{blood meals per patch, per day}

\item{pars}{a \link{list}}
}
\value{
fqZ, a numeric \link{vector}
}
\description{
Implements \link{fqZ_local} for the null model (do nothing)
}
