% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-interface.R
\name{fqZ_local}
\alias{fqZ_local}
\title{Correct fqZ for models with visitors}
\usage{
fqZ_local(fqZ, pars)
}
\arguments{
\item{fqZ}{blood meals per patch, per day}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Ipar}.
}
