% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inits}
\alias{get_inits}
\title{Get the initial values as a vector}
\usage{
get_inits(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Get the initial values as a vector
}
