% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{get_inits_H.dynamic}
\alias{get_inits_H.dynamic}
\title{Return initial values as a vector}
\usage{
\method{get_inits_H}{dynamic}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
