% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{get_inits_H.static}
\alias{get_inits_H.static}
\title{Return initial values as a vector}
\usage{
\method{get_inits_H}{static}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
