% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{get_inits_L.basic}
\alias{get_inits_L.basic}
\title{Return initial values as a vector}
\usage{
\method{get_inits_L}{basic}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{get_inits_L} for the GeRM model.
}
