% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{get_inits_MYZ.GeRM}
\alias{get_inits_MYZ.GeRM}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{GeRM}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{get_inits_MYZ} for the GeRM model.
}
