% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{get_inits_MYZ.Gtrace}
\alias{get_inits_MYZ.Gtrace}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{Gtrace}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_inits_MYZ} for the GeRM model.
}
