% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{get_inits_MYZ}
\alias{get_inits_MYZ}
\title{Return initial values as a vector}
\usage{
get_inits_MYZ(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
