% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{get_inits_MYZ.RM}
\alias{get_inits_MYZ.RM}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{RM}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{get_inits_MYZ} for the RM model.
}
