% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{get_inits_MYZ.RM_dde}
\alias{get_inits_MYZ.RM_dde}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{RM_dde}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_inits_MYZ} for the RM model.
}
