% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{get_inits_X}
\alias{get_inits_X}
\title{Return initial values as a vector}
\usage{
get_inits_X(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
