% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{get_inits_X.trace}
\alias{get_inits_X.trace}
\title{Return initial values as a vector}
\usage{
\method{get_inits_X}{trace}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
