% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-forced.R
\name{HabitatAvailability.forced}
\alias{HabitatAvailability.forced}
\title{Compute total availability of aquatic habitats}
\usage{
\method{HabitatAvailability}{forced}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of aquatic habitats
}
