% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-null.R
\name{HabitatAvailability.null}
\alias{HabitatAvailability.null}
\title{Compute total availability of aquatic habitats}
\usage{
\method{HabitatAvailability}{null}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of aquatic habitats for the null model (do nothing)
}
