% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{HabitatAvailability}
\alias{HabitatAvailability}
\title{Update the availability of aquatic habitats}
\usage{
HabitatAvailability(pars)
}
\arguments{
\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$RESOURCES}.
}
