% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics.null}
\alias{HabitatDynamics.null}
\title{Set the values of exogenous variables describing habitat dynamics}
\usage{
\method{HabitatDynamics}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{HabitatDynamics} for the null model of habitat_dynamics (do nothing)
}
