% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics}
\alias{HabitatDynamics}
\title{Set the values of exogenous variables describing habitat_dynamics}
\usage{
HabitatDynamics(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$HABITAT_DYNAMICS}.
}
