% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics.static}
\alias{HabitatDynamics.static}
\title{Set the values of habitat search weights and other exogenous variables describing habitat_dynamics}
\usage{
\method{HabitatDynamics}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{HabitatDynamics} for the static model of habitat_dynamics (do nothing)
}
