% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{HTC.hMoI}
\alias{HTC.hMoI}
\title{Compute the HTC for the hMoI model}
\usage{
\method{HTC}{hMoI}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the hMoI model with demography.
}
