% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{HTC}
\alias{HTC}
\title{Compute the human transmitting capacity}
\usage{
HTC(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
