% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{HTC.SIP}
\alias{HTC.SIP}
\title{Compute the HTC for the SIP model}
\usage{
\method{HTC}{SIP}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIP model with demography.
}
