% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{HTC.SIS}
\alias{HTC.SIS}
\title{Compute the HTC for the SIS model}
\usage{
\method{HTC}{SIS}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIS model with demography.
}
