% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanbehavior.R
\name{HumanBehavior.null}
\alias{HumanBehavior.null}
\title{Simulate no human behaviors}
\usage{
\method{HumanBehavior}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{HumanBehavior} for the null model (do nothing)
}
