% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanbehavior.R
\name{HumanBehavior}
\alias{HumanBehavior}
\title{Simulate Human Behavior}
\usage{
HumanBehavior(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$HBpar}.
}
