% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-null.R
\name{Hydrology.null}
\alias{Hydrology.null}
\title{Modify parameters due to exogenous forcing}
\usage{
\method{Hydrology}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{Hydrology} for the null model of hydrology (do nothing)
}
