% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{Import.basic}
\alias{Import.basic}
\title{Importation, the basic model}
\usage{
\method{Import}{basic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{Import} for the basic model of importation (do nothing)
}
