% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-null.R
\name{Import.null}
\alias{Import.null}
\title{Importation, the null model}
\usage{
\method{Import}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{Import} for the null model of importation (do nothing)
}
