% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-interface.R
\name{Import}
\alias{Import}
\title{Importation through visitors}
\usage{
Import(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Ipar}.
}
