% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-null.R
\name{IRS.null}
\alias{IRS.null}
\title{Modify baseline values due to IRS}
\usage{
\method{IRS}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a \link{list}
}
\description{
Implements \link{IRS} for the null model of IRS (do nothing)
}
