% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{IRS}
\alias{IRS}
\title{Modify baseline bionomic values due to IRS}
\usage{
IRS(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$IRSpar}.
}
