% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{IRS_EffectSizes}
\alias{IRS_EffectSizes}
\title{Model IRS effect sizes}
\usage{
IRS_EffectSizes(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$IRS}.
}
