% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-null.R
\name{IRSeffects.null}
\alias{IRSeffects.null}
\title{Model the effects of IRS}
\usage{
\method{IRSeffects}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$IRS}.
}
