% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-null.R
\name{IRSeffectSizes.null}
\alias{IRSeffectSizes.null}
\title{Model IRS effect sizes}
\usage{
\method{IRSeffectSizes}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$IRS}.
}
