% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{kappa_local.basic}
\alias{kappa_local.basic}
\title{Parasite / pathogen importation, the basic model}
\usage{
\method{kappa_local}{basic}(kappa, pars)
}
\arguments{
\item{kappa}{the net infectiousness among local strata}

\item{pars}{a \link{list}}
}
\value{
kappa, a numeric \link{vector}
}
\description{
Implements \link{kappa_local} for the basic model (do nothing)
}
