% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-null.R
\name{kappa_local.null}
\alias{kappa_local.null}
\title{Parasite / pathogen importation, the null model}
\usage{
\method{kappa_local}{null}(kappa, pars)
}
\arguments{
\item{kappa}{the net infectiousness among local strata}

\item{pars}{a \link{list}}
}
\value{
kappa, a numeric \link{vector}
}
\description{
Implements \link{kappa_local} for the null model (do nothing)
}
