% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-interface.R
\name{kappa_local}
\alias{kappa_local}
\title{Correct kappa for models with visitors}
\usage{
kappa_local(kappa, pars)
}
\arguments{
\item{kappa}{the net infectiousness among local strata}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Ipar}.
}
