% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-basic.R
\name{kappa_with_visitors.basic}
\alias{kappa_with_visitors.basic}
\title{kappa with visitors, the basic model}
\usage{
\method{kappa_with_visitors}{basic}(kappa, pars)
}
\arguments{
\item{kappa}{computed without visitors}

\item{pars}{a \link{list}}
}
\value{
a numeric \link{vector}
}
\description{
Implements \link{kappa_with_visitors} for the basic model of vector control (do nothing)
}
