% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-null.R
\name{kappa_with_visitors.null}
\alias{kappa_with_visitors.null}
\title{kappa with visitors, the null model}
\usage{
\method{kappa_with_visitors}{null}(kappa, pars)
}
\arguments{
\item{kappa}{computed without visitors}

\item{pars}{a \link{list}}
}
\value{
a numeric \link{vector}
}
\description{
Implements \link{kappa_with_visitors} for the null model of vector control (do nothing)
}
