% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-interface.R
\name{kappa_with_visitors}
\alias{kappa_with_visitors}
\title{Compute kappa with visitors}
\usage{
kappa_with_visitors(kappa, pars)
}
\arguments{
\item{kappa, }{computed without visitors}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$MIpar}.
}
