% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_to_inits}
\alias{last_to_inits}
\title{Set the initial values to the last values of the last simulation}
\usage{
last_to_inits(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
y a \link{numeric} vector
}
\description{
Set the initial values to the last values of the last simulation
}
