% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{LayEggs}
\alias{LayEggs}
\title{Compute eggs laid}
\usage{
LayEggs(t, y, pars, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Computes eggs laid for the s^th species
}
