% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{LBionomics.basic}
\alias{LBionomics.basic}
\title{Reset aquatic parameters to baseline}
\usage{
\method{LBionomics}{basic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{LBionomics} for the RM model
}
