% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-basic.R
\name{loc_fqZ.basic}
\alias{loc_fqZ.basic}
\title{fqZ with visitors, the basic model}
\usage{
\method{loc_fqZ}{basic}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a numeric \link{vector}
}
\description{
Implements \link{loc_fqZ} for the basic model of vector control (do nothing)
}
