% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-null.R
\name{loc_fqZ.null}
\alias{loc_fqZ.null}
\title{fqZ with visitors, the null model}
\usage{
\method{loc_fqZ}{null}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a numeric \link{vector}
}
\description{
Implements \link{loc_fqZ} for the null model of vector control (do nothing)
}
