% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-interface.R
\name{loc_fqZ}
\alias{loc_fqZ}
\title{Compute the local fraction, with visitors}
\usage{
loc_fqZ(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$MIpar}.
}
