% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fraction.R
\name{LocalFrac}
\alias{LocalFrac}
\title{Compute the local fraction}
\usage{
LocalFrac(pars)
}
\arguments{
\item{pars, }{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$BFpar}
}
