% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fraction.R
\name{LocalFrac.static}
\alias{LocalFrac.static}
\title{Compute transmission terms dynamically, no update required}
\usage{
\method{LocalFrac}{static}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$BFpar}
}
