% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-null.R
\name{LSM.null}
\alias{LSM.null}
\title{Modify baseline values due to LSM}
\usage{
\method{LSM}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a \link{list}
}
\description{
Implements \link{LSM} for the null model of LSM (do nothing)
}
