% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-interface.R
\name{LSM}
\alias{LSM}
\title{Modify values due to LSM}
\usage{
LSM(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSMpar}.
}
