% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-null.R
\name{LSM_Effects.null}
\alias{LSM_Effects.null}
\title{Modify effects of LSM, the null model}
\usage{
\method{LSM_Effects}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSM}
}
