% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-interface.R
\name{LSM_EffectSizes}
\alias{LSM_EffectSizes}
\title{Compute effect sizes of LSM}
\usage{
LSM_EffectSizes(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{y}{the state of the system}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSM}.
}
