% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-null.R
\name{LSMeffects.null}
\alias{LSMeffects.null}
\title{Modify effects of LSM, the null model}
\usage{
\method{LSMeffects}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{the state of the system}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSM}
}
