% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-interface.R
\name{LSMeffects}
\alias{LSMeffects}
\title{Modify effects of LSM}
\usage{
LSMeffects(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{the state of the system}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSM}.
}
