% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-dynamic.R
\name{make_beta.dynamic}
\alias{make_beta.dynamic}
\title{Biting distribution matrix}
\usage{
\method{make_beta}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
pars, a \link{list}
}
\description{
Implements \link{make_beta} for dynamic denominators

Compute beta, a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
