% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-interface.R
\name{make_beta}
\alias{make_beta}
\title{Call compute_beta as needed}
\usage{
make_beta(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
pars, a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Hpar}
}
