% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-static.R
\name{make_beta.static}
\alias{make_beta.static}
\title{Biting distribution matrix}
\usage{
\method{make_beta}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
pars, a \link{list}
}
\description{
Implements \link{make_beta} for static denominators
}
