% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-dynamic.R
\name{make_beta_lag.dynamic}
\alias{make_beta_lag.dynamic}
\title{Lagged biting distribution matrix}
\usage{
\method{make_beta_lag}{dynamic}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{is the lag value}
}
\value{
pars, a \link{list}
}
\description{
Implements \link{make_beta} for dynamic denominators

Compute beta, a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
