% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-interface.R
\name{make_beta_lag}
\alias{make_beta_lag}
\title{Call compute_beta with lagged variables, as needed}
\usage{
make_beta_lag(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{is the lag value}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$Hpar}
}
