% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixing-static.R
\name{make_beta_lag.static}
\alias{make_beta_lag.static}
\title{Biting distribution matrix}
\usage{
\method{make_beta_lag}{static}(t, y, pars, lag)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{lag}{is the lag value}
}
\value{
pars, a \link{list}
}
\description{
Implements \link{make_beta_lag} for static denominators
}
