% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{make_calD}
\alias{make_calD}
\title{Make the demography matrix \eqn{\mathcal{D}}}
\usage{
make_calD(d, m, b = NULL)
}
\arguments{
\item{d}{a vector of death rates (should have length equal to \code{nStrata})}

\item{m}{a vector of ageing rates (should have length equal to \code{nStrata-1})}

\item{b}{either \code{NULL} or a vector of birth rates (in which case should have length equal to \code{nStrata})}
}
\value{
a square \link{numeric} matrix of dimensions equal to the length of the input
argument \code{d}
}
\description{
Make the demography matrix \eqn{\mathcal{D}}, which multiplies a
state vector \code{Y} on the left to return the derivatives of the state with respect
to demographic dynamics. Please note that while lengths of arguments are given
below, the function does no argument checking as it is calculated internally
assuming correct inputs.
}
