% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{make_calK.as_matrix}
\alias{make_calK.as_matrix}
\title{Pass a pre-configured calK}
\usage{
\method{make_calK}{as_matrix}(nPatches, calK, opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{calK}{a matrix or a string}

\item{opts}{a list of options to configure calK}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_calK} for as_matrix
}
