% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{make_calK}
\alias{make_calK}
\title{Make a mosquito dispersal matrix, called calK}
\usage{
make_calK(nPatches, calK, opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{calK}{a matrix or a string}

\item{opts}{a list of options to configure calK}
}
\value{
a \link{matrix}
}
\description{
Make a mosquito dispersal matrix, called calK
}
