% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{make_calK.xy}
\alias{make_calK.xy}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
\method{make_calK}{xy}(nPatches, calK = "xy", opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{calK}{a matrix or a string}

\item{opts}{a list of options to configure calK}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_calK} for kernels
}
