% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{make_calK_xy}
\alias{make_calK_xy}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
make_calK_xy(xy, ker)
}
\arguments{
\item{xy}{is a vector of the xy-coordinates of patch locations}

\item{ker}{is a function that weights putative locations by distance}
}
\description{
Develop a mosquito dispersal matrix from a kernel and xy-coordinates
}
