% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_calN}
\alias{make_calN}
\title{Make the habitat membership matrix, calN}
\usage{
make_calN(nPatches, membership)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{membership}{is a vector describing the patch where each habitat is found}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nHabitats}
}
\description{
Make the habitat membership matrix, calN
}
