% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_calU}
\alias{make_calU}
\title{Make the egg distribution matrix, calU}
\usage{
make_calU(calN, searchQ = 1)
}
\arguments{
\item{calN}{the habitat membership matrix}

\item{searchQ}{the habitat search weights}
}
\value{
a \link{matrix} of dimensions \code{nHabitats} by \code{nPatches}
}
\description{
Make the egg distribution matrix, calU
}
