% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{make_calU_s}
\alias{make_calU_s}
\title{Make the egg distribution matrix, calU}
\usage{
make_calU_s(pars, s)
}
\arguments{
\item{pars}{the model object}

\item{s}{vector species index}
}
\value{
\link{list}
}
\description{
Make the egg distribution matrix, calU
}
