% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{make_Hpar_trace}
\alias{make_Hpar_trace}
\title{Make parameters for human null model}
\usage{
make_Hpar_trace(nPatches, Hopts = list())
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{Hopts}{a \link{list} that could overwrite defaults}
}
\value{
a \link{list}
}
\description{
Make parameters for human null model
}
