% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{make_index_H}
\alias{make_index_H}
\title{Add indices for human population denominators to parameter list}
\usage{
make_index_H(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
the modified parameter \link{list}
}
\description{
This method dispatches on the type of \code{pars$Hpar}.
}
