% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{make_index_H.trace}
\alias{make_index_H.trace}
\title{Add indices for human population denominators to parameter list}
\usage{
\method{make_index_H}{trace}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
the modified parameter \link{list}
}
\description{
Implements \link{make_index_H} for forced (trace) model.
}
