% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_index_L}
\alias{make_index_L}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
make_index_L(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
the modified parameter [list]
}
\description{
This method dispatches on the type of `pars$Lpar`. Adds field `L_ix`
to parameter list.
}
