% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{make_index_L.trace}
\alias{make_index_L.trace}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_index_L}{trace}(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
the modified parameter [list]
}
\description{
Implements [make_index_L] for trace (forced emergence) model.
}
