% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_index_MYZ.GeRM}
\alias{make_index_MYZ.GeRM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_index_MYZ}{GeRM}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_index_MYZ} for the generalized GeRM model.
}
