% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{make_index_MYZ}
\alias{make_index_MYZ}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
make_index_MYZ(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
