% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_index_MYZ.RM}
\alias{make_index_MYZ.RM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_index_MYZ}{RM}(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
the modified parameter [list]
}
\description{
Implements [make_index_MYZ] for the generalized RM model.
}
