% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_index_X.SIP}
\alias{make_index_X.SIP}
\title{Add indices for human population to parameter list}
\usage{
\method{make_index_X}{SIP}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
the modified parameter \link{list}
}
\description{
Implements \link{make_index_X} for the SIP model.
}
