% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_index_X.SIS}
\alias{make_index_X.SIS}
\title{Add indices for human population to parameter list}
\usage{
\method{make_index_X}{SIS}(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
the modified parameter [list]
}
\description{
Implements [make_index_X] for the SIS model.
}
