% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_indices}
\alias{make_indices}
\title{Set indices for generalized spatial model}
\usage{
make_indices(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
modified parameters \link{list}
}
\description{
Set indices for generalized spatial model
}
