% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-basic.R
\name{make_indices_H.basic}
\alias{make_indices_H.basic}
\title{Add indices for human population denominators to parameter list}
\usage{
\method{make_indices_H}{basic}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_indices_H} for forced (basic) model.
}
