% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{make_indices_H.dynamic}
\alias{make_indices_H.dynamic}
\title{Add indices for human population denominators to parameter list}
\usage{
\method{make_indices_H}{dynamic}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_H} for the dynamic demography model.
}
