% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-null.R
\name{make_indices_H.null}
\alias{make_indices_H.null}
\title{Add indices for human population denominators to parameter list}
\usage{
\method{make_indices_H}{null}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_indices_H} for null model.
}
