% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{make_indices_H}
\alias{make_indices_H}
\title{Add indices for human population denominators to parameter list}
\usage{
make_indices_H(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Hpar}.
}
