% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{make_indices_H.static}
\alias{make_indices_H.static}
\title{Add indices for human population denominators to parameter list}
\usage{
\method{make_indices_H}{static}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_H} for static model.
}
