% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{make_indices_L.basic}
\alias{make_indices_L.basic}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_L}{basic}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_L} for basic competition model.
}
