% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{make_indices_MYZ.basicM}
\alias{make_indices_MYZ.basicM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{basicM}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for the basic M model.
}
