% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_indices_MYZ.GeRM_dde}
\alias{make_indices_MYZ.GeRM_dde}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{GeRM_dde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for the GeRM model.
}
