% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_indices_MYZ.GeRM_ode}
\alias{make_indices_MYZ.GeRM_ode}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{GeRM_ode}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for the GeRM model.
}
