% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{make_indices_MYZ.Gtrace}
\alias{make_indices_MYZ.Gtrace}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{Gtrace}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for Gtrace (forced emergence) model.
}
