% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_indices_MYZ.RM}
\alias{make_indices_MYZ.RM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{RM}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for the generalized RM model.
}
