% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_indices_MYZ.RM_ode}
\alias{make_indices_MYZ.RM_ode}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{RM_ode}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for the RM model.
}
