% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{make_indices_MYZ.Ztrace}
\alias{make_indices_MYZ.Ztrace}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{Ztrace}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for Ztrace (forced emergence) model.
}
