% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{make_indices_X}
\alias{make_indices_X}
\title{Add indices for human population to parameter list}
\usage{
make_indices_X(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
