% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{make_indices_X.trace}
\alias{make_indices_X.trace}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{trace}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{make_indices_X} for the trace model.
}
