% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{make_inits_L_basic}
\alias{make_inits_L_basic}
\title{Make inits for basic competition aquatic mosquito model}
\usage{
make_inits_L_basic(pars, L0)
}
\arguments{
\item{pars}{an \link{environment}}

\item{L0}{initial conditions}
}
\value{
none
}
\description{
Make inits for basic competition aquatic mosquito model
}
