% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{make_inits_L_trace}
\alias{make_inits_L_trace}
\title{Make parameters for trace aquatic mosquito model}
\usage{
make_inits_L_trace(pars, L0 = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{L0}{is set to NULL for the trace model}
}
\value{
none
}
\description{
Make parameters for trace aquatic mosquito model
}
