% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_inits_MYZ_GeRM}
\alias{make_inits_MYZ_GeRM}
\title{Make inits for generalized GeRM ODE adult mosquito model}
\usage{
make_inits_MYZ_GeRM(pars, M0, G0, Y0, Z0, Upsilon0)
}
\arguments{
\item{pars}{an \link{environment}}

\item{M0}{total mosquito density at each patch}

\item{G0}{gravid mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate of gravid mosquitoes}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{tau}{length of extrinsic incubation period}
}
\value{
none
}
\description{
Make inits for generalized GeRM ODE adult mosquito model
}
