% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_inits_MYZ_GeRM_dde}
\alias{make_inits_MYZ_GeRM_dde}
\title{Make inits for GeRM adult mosquito model}
\usage{
make_inits_MYZ_GeRM_dde(pars, M0, G0, Y0, Z0, Upsilon0)
}
\arguments{
\item{pars}{an \link{environment}}

\item{M0}{total mosquito density at each patch}

\item{G0}{total parous mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}

\item{Upsilon0}{the initial values of Upsilon}
}
\value{
none
}
\description{
Make inits for GeRM adult mosquito model
}
