% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{make_inits_MYZ_Gtrace}
\alias{make_inits_MYZ_Gtrace}
\title{Make parameters for Gtrace aquatic mosquito model}
\usage{
make_inits_MYZ_Gtrace(pars, MYZ0 = NULL)
}
\arguments{
\item{pars}{an \link{environment}}

\item{MYZ0}{is set to NULL for the Gtrace model}
}
\value{
none
}
\description{
Make parameters for Gtrace aquatic mosquito model
}
