% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_inits_MYZ_RM}
\alias{make_inits_MYZ_RM}
\title{Make inits for generalized RM ODE adult mosquito model}
\usage{
make_inits_MYZ_RM(pars, M0, Y0, Z0, Upsilon0)
}
\arguments{
\item{pars}{an \link{environment}}

\item{M0}{total mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for generalized RM ODE adult mosquito model
}
