% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_inits_MYZ_RM_ode}
\alias{make_inits_MYZ_RM_ode}
\title{Make inits for RM adult mosquito model}
\usage{
make_inits_MYZ_RM_ode(pars, M0, P0, Y0, Z0)
}
\arguments{
\item{pars}{an \link{environment}}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for RM adult mosquito model
}
