% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{make_inits_MYZ_Ztrace}
\alias{make_inits_MYZ_Ztrace}
\title{Make parameters for Ztrace aquatic mosquito model}
\usage{
make_inits_MYZ_Ztrace(pars, MYZ0 = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZ0}{is set to NULL for the Ztrace model}
}
\value{
none
}
\description{
Make parameters for Ztrace aquatic mosquito model
}
