% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{make_inits_X_hMoI}
\alias{make_inits_X_hMoI}
\title{Make inits for hybrid MoI human model}
\usage{
make_inits_X_hMoI(pars, m10, m20)
}
\arguments{
\item{pars}{an \link{environment}}

\item{m10}{mean MoI among inapparent human infections}

\item{m20}{mean MoI among patent human infections}
}
\value{
none
}
\description{
MoI stands for Multiplicity of Infection, and refers to malarial superinfection.
}
