% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{make_inits_X_trace}
\alias{make_inits_X_trace}
\title{Make inits for trace human model}
\usage{
make_inits_X_trace(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make inits for trace human model
}
