% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{make_Linits_basic}
\alias{make_Linits_basic}
\title{Make inits for basic competition aquatic mosquito model}
\usage{
make_Linits_basic(nHabitats, Lopts = list(), L0 = 1)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{L0}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Make inits for basic competition aquatic mosquito model
}
