% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{make_Lpar_trace}
\alias{make_Lpar_trace}
\title{Make parameters for trace aquatic mosquito model}
\usage{
make_Lpar_trace(nHabitats, Lopts = list(), Lambda = 1000, Lt = NULL)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{Lambda}{vector of mean emergence rates from each aquatic habitat}

\item{Lt}{is a \link{function} of the form Lt(t,pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for trace aquatic mosquito model
}
