% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{make_MYZinits_basicM}
\alias{make_MYZinits_basicM}
\title{Make inits for basicM adult mosquito model}
\usage{
make_MYZinits_basicM(pars, MYZopts = list(), M0 = 5, P0 = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for basicM adult mosquito model
}
