% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_MYZinits_GeRM}
\alias{make_MYZinits_GeRM}
\title{Make inits for GeRM adult mosquito model}
\usage{
make_MYZinits_GeRM(pars, MYZopts = list(), M0 = 5, G0 = 1, Y0 = 1, Z0 = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZopts}{a \link{list} that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{G0}{total parous mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for GeRM adult mosquito model
}
