% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_MYZinits_GeRM_dde}
\alias{make_MYZinits_GeRM_dde}
\title{Make inits for GeRM adult mosquito model}
\usage{
make_MYZinits_GeRM_dde(
  nPatches,
  Upsilon,
  MYZopts = list(),
  M0 = 5,
  G0 = 1,
  Y0 = 1,
  Z0 = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{Upsilon}{a matrix describing survival and dispersal through the EIP}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{G0}{total parous mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for GeRM adult mosquito model
}
