% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_MYZinits_RM}
\alias{make_MYZinits_RM}
\title{Make inits for RM adult mosquito model}
\usage{
make_MYZinits_RM(pars, MYZopts = list(), M0 = 5, P0 = 1, Y0 = 1, Z0 = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RM adult mosquito model
}
