% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_MYZinits_RM_ode}
\alias{make_MYZinits_RM_ode}
\title{Make inits for RM adult mosquito model}
\usage{
make_MYZinits_RM_ode(
  nPatches,
  MYZopts = list(),
  M0 = 5,
  P0 = 1,
  Y0 = 1,
  Z0 = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RM adult mosquito model
}
