% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{make_MYZpar_basicM}
\alias{make_MYZpar_basicM}
\title{Make parameters for basicM ODE adult mosquito model}
\usage{
make_MYZpar_basicM(
  nPatches,
  MYZopts = list(),
  calK,
  g = 1/12,
  sigma = 1/8,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{the number of patches}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{f}{blood feeding rate}

\item{q}{human blood feeding fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a \link{list} with a configured MYZpar
}
\description{
Make parameters for basicM ODE adult mosquito model
}
