% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{make_MYZpar_Gtrace}
\alias{make_MYZpar_Gtrace}
\title{Make parameters for Gtrace aquatic mosquito model}
\usage{
make_MYZpar_Gtrace(pars, MYZopts, Gm = 1, Gf = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZopts}{a \link{list} to overwrite the defaults}

\item{Gm}{a vector of mean mosquito densities}

\item{Gf}{a \link{function} of the form Gf(t, pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for Gtrace aquatic mosquito model
}
