% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_MYZpar_RM}
\alias{make_MYZpar_RM}
\title{Make parameters for RM ODE adult mosquito model}
\usage{
make_MYZpar_RM(
  nPatches,
  MYZopts = list(),
  EIPmod,
  calK,
  g = 1/12,
  sigma = 1/8,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{EIPmod}{a \link{list} that defines the EIP model}

\item{calK}{a mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a \link{list}
}
\description{
Make parameters for RM ODE adult mosquito model
}
