% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{make_MYZpar_Ztrace}
\alias{make_MYZpar_Ztrace}
\title{Make parameters for Ztrace aquatic mosquito model}
\usage{
make_MYZpar_Ztrace(nPatches, MYZopts, f = 0.3, q = 0.95, Zm = 1, Zf = NULL)
}
\arguments{
\item{nPatches}{an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Zm}{a vector of mean mosquito densities}

\item{Zf}{a \link{function} of the form Zf(t, pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for Ztrace aquatic mosquito model
}
