% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodfeeding.R
\name{make_parameters_BF_static}
\alias{make_parameters_BF_static}
\title{Sets up BFpars for a model with static availability}
\usage{
make_parameters_BF_static(params)
}
\arguments{
\item{params}{is a list}
}
\value{
a \link{list} with BFpar added
}
\description{
Sets up BFpars for a model with static availability
}
