% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-basic.R
\name{make_parameters_demography_basic}
\alias{make_parameters_demography_basic}
\title{Make parameters for forced (basic) human demography model}
\usage{
make_parameters_demography_basic(
  pars,
  H,
  F_birth,
  birthrate,
  deathrate,
  birthsXstrata
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{H}{a function taking a single argument \code{t} and returning a vector of length
\code{nStrata}.}
}
\value{
none
}
\description{
Make parameters for forced (basic) human demography model
}
