% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{make_parameters_demography_dynamic}
\alias{make_parameters_demography_dynamic}
\title{Make parameters for forced (dynamic) human demography model}
\usage{
make_parameters_demography_dynamic(
  pars,
  H,
  residence,
  searchWts,
  TaR,
  birthFpars,
  Hmatrix,
  birthsXstrata
)
}
\arguments{
\item{pars}{a \link{list}}

\item{H}{a function taking a single argument \code{t} and returning a vector of length
\code{nStrata}.}

\item{residence}{is a vector describing patch residency}

\item{searchWts}{is a vector describing blood feeding search weights}

\item{TaR}{is a matrix describing time spent among patches}

\item{birthFpars}{setup to dispatch and compute \code{F_birth}}

\item{Hmatrix}{does a set of state transitions}

\item{birthsXstrata}{distributes births to the youngest strata}
}
\value{
none
}
\description{
Make parameters for forced (dynamic) human demography model
}
