% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{make_parameters_demography_static}
\alias{make_parameters_demography_static}
\title{Make parameters for static human demography model}
\usage{
make_parameters_demography_static(pars, H, membershipH, searchWtsH, TimeSpent)
}
\arguments{
\item{pars}{an \link{environment}}

\item{H}{size of human population in each strata}

\item{membershipH}{is a vector describing patch residency}

\item{searchWtsH}{is a vector describing blood feeding search weights}

\item{TimeSpent}{is a matrix describing time spent among patches}
}
\value{
none
}
\description{
Make parameters for static human demography model
}
