% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-trace.R
\name{make_parameters_demography_trace}
\alias{make_parameters_demography_trace}
\title{Make parameters for forced (trace) human demography model}
\usage{
make_parameters_demography_trace(pars, H)
}
\arguments{
\item{pars}{an \link{environment}}

\item{H}{a function taking a single argument \code{t} and returning a vector of length
\code{nStrata}.}
}
\value{
none
}
\description{
Make parameters for forced (trace) human demography model
}
