% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{make_parameters_exposure_pois}
\alias{make_parameters_exposure_pois}
\title{Make parameters for the null model of exposure}
\usage{
make_parameters_exposure_pois(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model of exposure
}
