% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{make_parameters_import_basic}
\alias{make_parameters_import_basic}
\title{Make parameters for the basic model for parasite / pathogen importation}
\usage{
make_parameters_import_basic(pars, local_frac = 1, x_visitors = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{local_frac}{the fraction of bites taken on residents}

\item{x_visitors}{the probability a bite on a visitor infects a mosquito}
}
\value{
none
}
\description{
Make parameters for the basic model for parasite / pathogen importation
}
