% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-null.R
\name{make_parameters_import_null}
\alias{make_parameters_import_null}
\title{Make parameters for the null model for parasite / pathogen importation}
\usage{
make_parameters_import_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model for parasite / pathogen importation
}
