% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{make_parameters_itn_null}
\alias{make_parameters_itn_null}
\title{Make parameters for the null model of bed nets (do nothing)}
\usage{
make_parameters_itn_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Make parameters for the null model of bed nets (do nothing)
}
