% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-basic.R
\name{make_parameters_mi_basic}
\alias{make_parameters_mi_basic}
\title{Make parameters for the basic model for malaria importation}
\usage{
make_parameters_mi_basic(pars, local_fraction, X_delta)
}
\arguments{
\item{pars}{a \link{list}}

\item{local_fraction}{a numeric \link{vector} describing the fraction local}

\item{X_delta}{a numeric \link{vector} describing the density of infectious visitors}
}
\value{
none
}
\description{
Make parameters for the basic model for malaria importation
}
