% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-null.R
\name{make_parameters_mi_null}
\alias{make_parameters_mi_null}
\title{Make parameters for the null model for malaria importation}
\usage{
make_parameters_mi_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model for malaria importation
}
