% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_parameters_MYZ_GeRM}
\alias{make_parameters_MYZ_GeRM}
\title{Make parameters for generalized GeRM ODE adult mosquito model}
\usage{
make_parameters_MYZ_GeRM(
  pars,
  g,
  sigma,
  f,
  q,
  nu,
  eggsPerBatch,
  tau,
  calK,
  solve_as = "dde"
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate of gravid mosquitoes}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{tau}{length of extrinsic incubation period}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{solve_as}{is a switch: either 'ode' or 'dde'}
}
\value{
none
}
\description{
Make parameters for generalized GeRM ODE adult mosquito model
}
