% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_parameters_MYZ_GeRM_dde}
\alias{make_parameters_MYZ_GeRM_dde}
\title{Make parameters for generalized GeRM DDE adult mosquito model}
\usage{
make_parameters_MYZ_GeRM_dde(
  pars,
  g,
  sigma,
  calK,
  f,
  q,
  nu,
  eggsPerBatch,
  tau,
  M0,
  G0,
  Y0,
  Z0
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate of gravid mosquitoes}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{tau}{length of extrinsic incubation period}

\item{M0}{total mosquito density at each patch}

\item{G0}{gravid mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
none
}
\description{
Make parameters for generalized GeRM DDE adult mosquito model
}
