% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_parameters_MYZ_GeRM_exo}
\alias{make_parameters_MYZ_GeRM_exo}
\title{Make parameters for generalized GeRM ODE adult mosquito model}
\usage{
make_parameters_MYZ_GeRM_exo(
  pars,
  F_g,
  F_sigma,
  F_f,
  F_q,
  F_nu,
  eggsPerBatch,
  tau,
  calK,
  solve_as = "dde"
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{F_g}{function to compute the mosquito mortality rate}

\item{F_sigma}{function to compute the emigration rate}

\item{F_f}{function to compute the feeding rate}

\item{F_q}{function to compute the human blood fraction}

\item{F_nu}{function to compute oviposition rate of gravid mosquitoes}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{tau}{length of extrinsic incubation period}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{solve_as}{is a switch: either 'ode' or 'dde'}
}
\value{
none
}
\description{
Make parameters for generalized GeRM ODE adult mosquito model
}
