% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{make_parameters_MYZ_Gtrace}
\alias{make_parameters_MYZ_Gtrace}
\title{Make parameters for Gtrace aquatic mosquito model}
\usage{
make_parameters_MYZ_Gtrace(pars, Gm, Gf)
}
\arguments{
\item{pars}{an \link{environment}}

\item{Gm}{a vector of mean mosquito densities}

\item{Gf}{a \link{function} of the form Gf(t, pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for Gtrace aquatic mosquito model
}
