% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_parameters_MYZ_RM_ode}
\alias{make_parameters_MYZ_RM_ode}
\title{Make parameters for generalized RM ODE adult mosquito model}
\usage{
make_parameters_MYZ_RM_ode(
  Omega,
  OmegaEIP,
  f,
  q,
  nu,
  eggsPerBatch,
  M0,
  G0,
  Y0,
  Z0
)
}
\arguments{
\item{Omega}{mosquito demography matrix}

\item{OmegaEIP}{mosquito demography matrix through the EIP}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate of gravid mosquitoes}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{M0}{total mosquito density at each patch}

\item{G0}{gravid mosquito density at each patch}

\item{Y0}{infected mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list} with classes \code{RM}, \code{RM_ode}.
}
\description{
Make parameters for generalized RM ODE adult mosquito model
}
