% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorcontrol-control.R
\name{make_parameters_vc_control}
\alias{make_parameters_vc_control}
\title{Make parameters for the control model of vector control (do nothing)}
\usage{
make_parameters_vc_control(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Make parameters for the control model of vector control (do nothing)
}
