% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorcontrol-null.R
\name{make_parameters_vc_null}
\alias{make_parameters_vc_null}
\title{Make parameters for the null model of vector control (do nothing)}
\usage{
make_parameters_vc_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model of vector control (do nothing)
}
