% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-null.R
\name{make_parameters_weather_null}
\alias{make_parameters_weather_null}
\title{Make parameters for the null model for weather (do nothing)}
\usage{
make_parameters_weather_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model for weather (do nothing)
}
