% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_parameters_X_SIP}
\alias{make_parameters_X_SIP}
\title{Make parameters for SIP human model}
\usage{
make_parameters_X_SIP(b, c, r, rho, eta)
}
\arguments{
\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}

\item{rho}{probability of successful treatment upon infection}

\item{eta}{prophylaxis waning rate}
}
\value{
a \link{list} with class \code{SIP}.
}
\description{
Make parameters for SIP human model
}
