% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{make_parameters_X_trace}
\alias{make_parameters_X_trace}
\title{Make parameters for trace human model}
\usage{
make_parameters_X_trace(pars, kappa)
}
\arguments{
\item{pars}{an \link{list}}

\item{kappa}{net infectiousness}
}
\value{
a \link{list}
}
\description{
Make parameters for trace human model
}
