% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeatrisk.R
\name{make_TaR.as_matrix}
\alias{make_TaR.as_matrix}
\title{Pass a pre-configured TaR}
\usage{
\method{make_TaR}{as_matrix}(nPatches, residence, TaR, opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the residence vector}

\item{TaR}{a matrix; or a string}

\item{opts}{a list of options to configure TaR}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_TaR} for as_matrix
}
