% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_at_risk.R
\name{make_TaR.athome}
\alias{make_TaR.athome}
\title{Make a mosquito dispersal matrix, called TaR with a here / away}
\usage{
\method{make_TaR}{athome}(nPatches, residence, TaR = "athome", opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the residence vector}

\item{TaR}{a matrix; or a string}

\item{opts}{a list of options to configure TaR}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_TaR} for as_matrix
}
