% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{make_TaR}
\alias{make_TaR}
\title{Make TaR}
\usage{
make_TaR(t, pars, i, s)
}
\arguments{
\item{t}{the time}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{s}{the vector species index}
}
\value{
none
}
\description{
Make a time at risk matrix (TaR) from a time spent matrix and a circadian function
}
