% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_at_risk.R
\name{make_TaR.xy}
\alias{make_TaR.xy}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
\method{make_TaR}{xy}(nPatches, residence, TaR = "xy", opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the residence vector}

\item{TaR}{a matrix; or a string}

\item{opts}{a list of options to configure TaR}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_TaR} for kernels
}
