% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeatrisk.R
\name{make_TaR_athome}
\alias{make_TaR_athome}
\title{Make a mosquito dispersal matrix, called TaR}
\usage{
make_TaR_athome(nPatches, residence, opts, atHome = 0.95, travel = 0.01)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the home patch for each stratum}

\item{opts}{is a set of options that overwrites the defaults}

\item{atHome}{is the fraction of time spent at home}

\item{travel}{is the fraction of time spent traveling}
}
\value{
a \link{matrix}
}
\description{
Make a mosquito dispersal matrix, called TaR
}
