% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeatrisk.R
\name{make_TaR_xy}
\alias{make_TaR_xy}
\title{Make a mosquito dispersal matrix, called TaR}
\usage{
make_TaR_xy(xy, residence, kern, stay, travel)
}
\arguments{
\item{xy}{is the xy-locations of the patches}

\item{residence}{is the home patch for each stratum}

\item{kern}{is a function that gives weight by distance}

\item{stay}{is the fraction of time spent at home}

\item{travel}{is the fraction of time spent traveling}
}
\value{
a \link{matrix}
}
\description{
Make a mosquito dispersal matrix, called TaR
}
