% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_TimeSpent.as_matrix}
\alias{make_TimeSpent.as_matrix}
\title{Pass a pre-configured TimeSpent}
\usage{
\method{make_TimeSpent}{as_matrix}(pars, i, TimeSpent, opts = list())
}
\arguments{
\item{pars}{an \link{list}}

\item{i}{the host species index}

\item{TimeSpent}{a matrix; or a string}

\item{opts}{a list of options to configure TimeSpent}
}
\value{
a \link{list}
}
\description{
Implements \link{make_TimeSpent} for as_matrix
}
