% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_TimeSpent.athome}
\alias{make_TimeSpent.athome}
\title{Make a mosquito dispersal matrix, called TimeSpent with a here / away}
\usage{
\method{make_TimeSpent}{athome}(pars, i, TimeSpent = "athome", opts = list())
}
\arguments{
\item{pars}{an \link{list}}

\item{i}{the host species index}

\item{TimeSpent}{a matrix; or a string}

\item{opts}{a list of options to configure TimeSpent}
}
\value{
a \link{list}
}
\description{
Implements \link{make_TimeSpent} for as_matrix
}
