% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_TimeSpent}
\alias{make_TimeSpent}
\title{Make a time spent matrix, called TimeSpent}
\usage{
make_TimeSpent(pars, i, TimeSpent, opts = list())
}
\arguments{
\item{pars}{an \link{list}}

\item{i}{the host species index}

\item{TimeSpent}{a matrix; or a string}

\item{opts}{a list of options to configure TimeSpent}
}
\value{
a \link{list}
}
\description{
Make a time spent matrix, called TimeSpent
}
