% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_TimeSpent.xy}
\alias{make_TimeSpent.xy}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
\method{make_TimeSpent}{xy}(pars, i, TimeSpent = "xy", opts = list())
}
\arguments{
\item{pars}{an \link{list}}

\item{i}{the host species index}

\item{TimeSpent}{a matrix; or a string}

\item{opts}{a list of options to configure TimeSpent}
}
\value{
a \link{list}
}
\description{
Implements \link{make_TimeSpent} for kernels
}
