% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_TimeSpent_athome}
\alias{make_TimeSpent_athome}
\title{Make a mosquito dispersal matrix, called TimeSpent}
\usage{
make_TimeSpent_athome(
  nPatches,
  residence,
  opts = list(),
  atHome = 1,
  travel = 0
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the home patch for each stratum}

\item{opts}{is a set of options that overwrites the defaults}

\item{atHome}{is the fraction of time spent at home}

\item{travel}{is the fraction of time spent traveling}
}
\value{
a \link{matrix}
}
\description{
Make a mosquito dispersal matrix, called TimeSpent
}
