% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{make_Xinits_hMoI}
\alias{make_Xinits_hMoI}
\title{Make inits for hybrid MoI human model}
\usage{
make_Xinits_hMoI(pars, Xopts = list(), m10 = 2, m20 = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} that overwrites default values}

\item{m10}{mean MoI among inapparent human infections}

\item{m20}{mean MoI among patent human infections}
}
\value{
none
}
\description{
MoI stands for Multiplicity of Infection, and refers to malarial superinfection.
}
