% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_Xinits_SIS}
\alias{make_Xinits_SIS}
\title{Make initial values for the SIS human model, with defaults}
\usage{
make_Xinits_SIS(pars, Xopts = list(), S0 = NULL, I0 = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{S0}{the initial values of the parameter S}

\item{I0}{the initial values of the parameter I}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIS human model, with defaults
}
