% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{make_Xpar_hMoI}
\alias{make_Xpar_hMoI}
\title{Make parameters for hybrid MoI human model}
\usage{
make_Xpar_hMoI(
  pars,
  Xopts = list(),
  b = 0.55,
  r1 = 1/180,
  r2 = 1/70,
  c1 = 0.015,
  c2 = 0.15
)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} that overwrites default values}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r1}{recovery rate from inapparent infections}

\item{r2}{recovery rate from patent infections}

\item{c1}{transmission probability (efficiency) from inapparent human infections to mosquito}

\item{c2}{transmission probability (efficiency) from patent human infections to mosquito}
}
\value{
none
}
\description{
MoI stands for Multiplicity of Infection, and refers to malarial superinfection.
}
