% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_Xpar_SIP}
\alias{make_Xpar_SIP}
\title{Make parameters for SIP human model, with defaults}
\usage{
make_Xpar_SIP(
  pars,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  c = 0.15,
  rho = 0.1,
  eta = 1/25,
  xi = 1/365
)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{rho}{probability of successful treatment upon infection}

\item{eta}{prophylaxis waning rate}

\item{xi}{background treatment rate}
}
\value{
a \link{list}
}
\description{
Make parameters for SIP human model, with defaults
}
