% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{make_Xpar_trace}
\alias{make_Xpar_trace}
\title{Make parameters for human null model}
\usage{
make_Xpar_trace(pars, Xopts = list(), kappa = 0.1, Kf = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{kappa}{value}

\item{Kf}{a function}
}
\value{
a \link{list}
}
\description{
Make parameters for human null model
}
