% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-basic.R
\name{MalariaImportation.basic}
\alias{MalariaImportation.basic}
\title{Malaria importation, the basic model}
\usage{
\method{MalariaImportation}{basic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a named \link{list}
}
\description{
Implements \link{MalariaImportation} for the basic model of vector control (do nothing)
}
