% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-null.R
\name{MalariaImportation.null}
\alias{MalariaImportation.null}
\title{Malaria importation, the null model}
\usage{
\method{MalariaImportation}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{environment}}
}
\value{
a named \link{list}
}
\description{
Implements \link{MalariaImportation} for the null model of vector control (do nothing)
}
