% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_medical.R
\name{MassMedical.forced}
\alias{MassMedical.forced}
\title{Methods for mass medical}
\usage{
\method{MassMedical}{forced}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{MassMedical}
}
