% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_medical.R
\name{MassMedical}
\alias{MassMedical}
\title{Methods for mass medical interventions}
\usage{
MassMedical(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$MASS_MED}.
}
