% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{MBionomics.Gtrace}
\alias{MBionomics.Gtrace}
\title{Compute bloodfeeding and mortality rates}
\usage{
\method{MBionomics}{Gtrace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{MBionomics} for the Gtrace model.
}
