% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{MBionomics}
\alias{MBionomics}
\title{Set bloodfeeding and mortality rates to baseline}
\usage{
MBionomics(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}. It should
set the values of the bionomic parameters to baseline values.
}
