% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metric_calR}
\alias{metric_calR}
\title{Parasite Dispersal through one Parasite Generation (Humans)}
\usage{
metric_calR(b, beta, calV, W, D, H)
}
\arguments{
\item{b}{transmission efficiency from mosquitoes to humans}

\item{beta}{the biting distribution matrix}

\item{calV}{parasite dispersal by mosquitoes matrix (see \link{metric_calV})}

\item{W}{ambient human population at each patch}

\item{D}{human transmitting capacity}

\item{H}{human population size of each strata}
}
\value{
a numeric \link{matrix}
}
\description{
Computes a \code{n} by \code{n} matrix describing parasite dispersal from infecteds (columns)
to infectees (rows).
\deqn{\mathcal{R} =   b \beta \cdot {\cal V}  \cdot \mbox{diag}\left(W \right) \cdot \beta^T  \cdot \mbox{diag}\left(DH\right)}
}
