% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{MosquitoBehavior}
\alias{MosquitoBehavior}
\title{Compute bloodfeeding and mortality rates}
\usage{
MosquitoBehavior(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}. It should,
at a minimum return the values \code{f}, \code{q}, \code{g} (blood feeding rate, human feeding
proportion, and mortality rate) at the current time, although it may return
vectors of these values at multiple times for models with delay. These baseline
values will be modified by the vector control component. The return type is a
named list with those 3 values, and \code{f}  should have an \link{attr} labled \code{time}
giving the time(s) in the simulation that these bionomic values correspond to.
}
