% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{MosquitoBehavior.RM}
\alias{MosquitoBehavior.RM}
\title{Reset bloodfeeding and mortality rates to baseline}
\usage{
\method{MosquitoBehavior}{RM}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{MosquitoBehavior} for the RM model
}
