% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{OtherBloodHosts.static}
\alias{OtherBloodHosts.static}
\title{Set the values of exogenous variables describing other blood hosts}
\usage{
\method{OtherBloodHosts}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{OtherBloodHosts} for the static model of other_blood_hosts (do nothing)
}
