% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{OwnBedNet.null}
\alias{OwnBedNet.null}
\title{Bed net ownership}
\usage{
\method{OwnBedNet}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{OwnBedNet} for the null model of bed nets (do nothing)
}
