% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-interface.R
\name{OwnBedNet}
\alias{OwnBedNet}
\title{Model bed net loss, called from Control(VectorControl)}
\usage{
OwnBedNet(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$ITNown}.
}
