% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_deout}
\alias{parse_deout}
\title{Parse the output of an object returned by deSolve}
\usage{
parse_deout(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of orbits returned by deSolve}

\item{pars}{a \link{list}}
}
\value{
varslist a \link{list}
}
\description{
Parse the output of an object returned by deSolve
}
