% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{parse_deout_H.dynamic}
\alias{parse_deout_H.dynamic}
\title{Parse the output of deSolve and return H for models with dynamic demography}
\usage{
\method{parse_deout_H}{dynamic}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_deout_H} for models with dynamic demography
}
