% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{parse_deout_H}
\alias{parse_deout_H}
\title{Parse the output of deSolve and return the demography variables by name in a list}
\usage{
parse_deout_H(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\description{
This method dispatches on the type of \code{pars$Hpar}. Adds the variables
from the demography module to a list and returns it
}
