% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{parse_deout_H.static}
\alias{parse_deout_H.static}
\title{Parse the output of deSolve and return variables for models where H is a parameter}
\usage{
\method{parse_deout_H}{static}(varslist, deout, pars)
}
\arguments{
\item{varslist}{a \link{list} the object to be returned}

\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
none
}
\description{
Implements \link{parse_deout_H} for models with constant denominators
}
