% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{parse_deout_L.basic}
\alias{parse_deout_L.basic}
\title{Parse the variable names for the basic model}
\usage{
\method{parse_deout_L}{basic}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines the model}
}
\value{
\link{list}
}
\description{
Implements \link{parse_deout_L} for basic competition model.
}
