% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{parse_deout_L.trace}
\alias{parse_deout_L.trace}
\title{Parse the variable names for the trace model}
\usage{
\method{parse_deout_L}{trace}(varslist, deout, pars)
}
\arguments{
\item{varslist}{a \link{list} the object to be returned}

\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
varslist a \link{list}
}
\description{
Implements \link{parse_deout_L} for the trace model.
}
