% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{parse_deout_MYZ.basicM}
\alias{parse_deout_MYZ.basicM}
\title{Parse the output of deSolve and return variables for the basicM model}
\usage{
\method{parse_deout_MYZ}{basicM}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
none
}
\description{
Implements \link{parse_deout_MYZ} for the basicM model.
}
