% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{parse_deout_MYZ}
\alias{parse_deout_MYZ}
\title{Parse the output of deSolve and return the variables by name in a list}
\usage{
parse_deout_MYZ(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
It computes the variables by name and returns a named list.
}
