% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{parse_deout_MYZ.Ztrace}
\alias{parse_deout_MYZ.Ztrace}
\title{Parse the output of deSolve and return variables for the Ztrace model}
\usage{
\method{parse_deout_MYZ}{Ztrace}(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_deout_MYZ} for Ztrace
}
