% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{parse_deout_X.SIP}
\alias{parse_deout_X.SIP}
\title{Parse the output of deSolve and return variables for the SIP model}
\usage{
\method{parse_deout_X}{SIP}(varslist, deout, pars)
}
\arguments{
\item{varslist}{a \link{list} the object to be returned}

\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}
}
\value{
none
}
\description{
Implements \link{parse_deout_X} for the SIP model
}
