% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Rainfall.null}
\alias{Rainfall.null}
\title{Methods for exogenous variables describing rainfall}
\usage{
\method{Rainfall}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Rainfall} for the null model (no variables)
}
