% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Rainfall}
\alias{Rainfall}
\title{Methods for exogenous variables describing rainfall}
\usage{
Rainfall(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$RAINFALL}.
}
