% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Relative_Humidity.null}
\alias{Relative_Humidity.null}
\title{Methods for exogenous variables describing relative humidity}
\usage{
\method{Relative_Humidity}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Relative_Humidity} for the null model (no variables)
}
