% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Relative_Humidity}
\alias{Relative_Humidity}
\title{Methods for exogenous variables describing relative humidity}
\usage{
Relative_Humidity(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$HUMIDITY}.
}
