% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{ResourceAvailability}
\alias{ResourceAvailability}
\title{Set the values of exogenous variables describing available non-host resources}
\usage{
ResourceAvailability(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$RESpar}.
}
