% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{Resources.dynamic}
\alias{Resources.dynamic}
\title{Modify resources and resource availability dynamically}
\usage{
\method{Resources}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{Resources} for the null model of resources (do nothing)
}
