% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-forced.R
\name{Resources.forced}
\alias{Resources.forced}
\title{Methods for resources}
\usage{
\method{Resources}{forced}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Resources}
}
