% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{Resources.GeRM}
\alias{Resources.GeRM}
\title{Set the availability of resources}
\usage{
\method{Resources}{GeRM}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{Resources} for the GeRM model
}
