% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-null.R
\name{Resources.null}
\alias{Resources.null}
\title{Modify resources and resource availability}
\usage{
\method{Resources}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{Resources} for the null model of resources (do nothing)
}
