% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{Resources}
\alias{Resources}
\title{Set the values of exogenous variables describing available mosquito resources}
\usage{
Resources(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$RApar}.
}
