% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Resources.static}
\alias{Resources.static}
\title{Modify resources and resource availability}
\usage{
\method{Resources}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{Resources} for the null model of resources
}
