% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic.R
\name{setup_abiotic_forced}
\alias{setup_abiotic_forced}
\title{Make parameters for the null model of abiotic forcing (do nothing)}
\usage{
setup_abiotic_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the null model of abiotic forcing (do nothing)
}
