% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic.R
\name{setup_abiotic_null}
\alias{setup_abiotic_null}
\title{Set up the null model for exogenous forcing (do nothing)}
\usage{
setup_abiotic_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for exogenous forcing (do nothing)
}
