% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{setup_area_spray_null}
\alias{setup_area_spray_null}
\title{Set up the null model for area spraying (do nothing)}
\usage{
setup_area_spray_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for area spraying (do nothing)
}
