% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{setup_behavior.null}
\alias{setup_behavior.null}
\title{Setup behavior}
\usage{
\method{setup_behavior}{null}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Setup behavior
}
