% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{setup_BFpar_static}
\alias{setup_BFpar_static}
\title{Set up the blood feeding model}
\usage{
setup_BFpar_static(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This sets up a list that stores all the information required by the
blood feeding model, including information about humans (the residence vector, search
weights for blood feeding mosquitoes); the mosquitoes circadian pattern (F_circadian);
and derived structures, including the time spent (TimeSpent) and the time at risk (TaR)
matrices. This also sets up the objects required for blood feeding, including host availability
for each species (Wi) and for all hosts (W), other blood hosts (Other), and total blood
availability (B). Finally, it sets up the lists that will hold the mixing matrix (beta), and
the transmission terms describing the entomological inoculation rate (EIR and eir), and the force
of infection (FoI), and the net infectiousness (kappa and ni).
}
