% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{setup_bionomics_static}
\alias{setup_bionomics_static}
\title{Make parameters for the static model bionomics}
\usage{
setup_bionomics_static(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the static model bionomics
}
