% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{setup_BloodFeeding}
\alias{setup_BloodFeeding}
\title{Set up blood feeding}
\usage{
setup_BloodFeeding(
  pars,
  i,
  s = 1,
  BFopts = list(),
  residence = 1,
  searchWts = 1,
  F_circadian = NULL
)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{s}{the vector species index}

\item{BFopts}{a \link{list}}

\item{residence}{is the patch where each stratum resides}

\item{searchWts}{is the blood feeding search weight for each stratum}

\item{F_circadian}{is a function that computes relative mosquito blood feeding activity rates by time of day}
}
\value{
none
}
\description{
This sets up a list that stores all the information
}
