% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil.R
\name{setup_civil_null}
\alias{setup_civil_null}
\title{Set up the null model for exogenous forcing (do nothing)}
\usage{
setup_civil_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for exogenous forcing (do nothing)
}
