% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{setup_clinic_forced}
\alias{setup_clinic_forced}
\title{Set up a model for clinic}
\usage{
setup_clinic_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up a model for clinic
}
