% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{setup_clinic_null}
\alias{setup_clinic_null}
\title{Set up the null model for clinical distribution (do nothing)}
\usage{
setup_clinic_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for clinical distribution (do nothing)
}
