% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{setup_control.null}
\alias{setup_control.null}
\title{Setup control}
\usage{
\method{setup_control}{null}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Setup control
}
