% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{setup_control}
\alias{setup_control}
\title{Setup CONTROL with forcing}
\usage{
setup_control(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Setup CONTROL with forcing
}
