% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{setup_control_forced}
\alias{setup_control_forced}
\title{Set up a model with some control}
\usage{
setup_control_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up a model with some control
}
