% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{setup_development_null}
\alias{setup_development_null}
\title{Set up the null model for developments (do nothing)}
\usage{
setup_development_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for developments (do nothing)
}
