% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_EIP}
\alias{setup_EIP}
\title{Set up the static model for control forcing (do nothing)}
\usage{
setup_EIP(EIPname = "static", EIPopts = list())
}
\arguments{
\item{EIPname}{the class name of the function}

\item{EIPopts}{is a \link{list} that overwrites default options}
}
\value{
\link{list}
}
\description{
Set up the static model for control forcing (do nothing)
}
