% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_EIP.static}
\alias{setup_EIP.static}
\title{Set up a static model for the EIP}
\usage{
\method{setup_EIP}{static}(EIPname, EIPopts = list())
}
\arguments{
\item{EIPname}{the class name of the function}

\item{EIPopts}{is a \link{list} that overwrites default options}
}
\value{
\link{list}
}
\description{
Set up a static model for the EIP
}
