% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_eip_static}
\alias{setup_eip_static}
\title{Set up the static model for control forcing (do nothing)}
\usage{
setup_eip_static(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the static model for control forcing (do nothing)
}
