% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{setup_forcing_MYZ_GeRM_basic}
\alias{setup_forcing_MYZ_GeRM_basic}
\title{Set up a static exogenous forcing for the GeRM ODE adult mosquito model}
\usage{
setup_forcing_MYZ_GeRM_basic(pars, other, sugar, zeta)
}
\arguments{
\item{pars}{a \link{list}}

\item{other}{is the availability of other blood hosts}

\item{sugar}{is sugar availability}

\item{zeta}{is a shape parameter}
}
\value{
none
}
\description{
Set up a static exogenous forcing for the GeRM ODE adult mosquito model
}
