% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{setup_habitats_forced}
\alias{setup_habitats_forced}
\title{Set up forcing for the availability of habitats}
\usage{
setup_habitats_forced(
  pars,
  habitats_Opts = list(),
  habitats_scale = 0,
  habitats_t = NULL
)
}
\arguments{
\item{pars}{a \link{list}}

\item{habitats_Opts}{a list, overwrites default values}

\item{habitats_scale}{a non-negative numeric value to scale the mean availability of habitats}

\item{habitats_t}{a function that specifies the temporal pattern for the availability of habitats}
}
\value{
none
}
\description{
Set up forcing for the availability of habitats
}
