% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{setup_habitats_static}
\alias{setup_habitats_static}
\title{Set up the static model for habitat dynamics}
\usage{
setup_habitats_static(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Set up the static model for habitat dynamics
}
