% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{setup_Hpar}
\alias{setup_Hpar}
\title{A utility to set up Hpar}
\usage{
setup_Hpar(pars, HPop = 1000, residence = 1, searchWts = 1, Hopts = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{HPop}{a \link{numeric} vector of population densities}

\item{residence}{is the patch where each stratum resides}

\item{searchWts}{is the blood feeding search weight for each stratum}

\item{Hopts}{a \link{list} to overwrite default values}
}
\value{
a \link{list}
}
\description{
A utility to set up Hpar
}
