% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{setup_Hpar_static}
\alias{setup_Hpar_static}
\title{A utility to set up Hpar}
\usage{
setup_Hpar_static(pars, i, HPop = 1000)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{HPop}{a \link{numeric} vector of population densities}
}
\value{
a \link{list}
}
\description{
A utility to set up Hpar
}
