% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanbehavior.R
\name{setup_humanbehavior_null}
\alias{setup_humanbehavior_null}
\title{Make parameters for the null model for resource availability (do nothing)}
\usage{
setup_humanbehavior_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model for resource availability (do nothing)
}
