% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-null.R
\name{setup_hydrology_null}
\alias{setup_hydrology_null}
\title{Make parameters for the null model for hydrology (do nothing)}
\usage{
setup_hydrology_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model for hydrology (do nothing)
}
