% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{setup_L}
\alias{setup_L}
\title{A function to set up Lpar}
\usage{
setup_L(pars, Lname, membership = 1, searchQ = 1, Lopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Lname}{a \link{character} string}

\item{membership}{a \link{vector} of length \code{nHabitats}}

\item{searchQ}{a \link{vector} of length \code{nHabitats}}

\item{Lopts}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on \code{Lname}.
}
