% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{setup_L.trace}
\alias{setup_L.trace}
\title{Setup Lpar.trace}
\usage{
\method{setup_L}{trace}(pars, Lname, membership = 1, searchQ = 1, Lopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Lname}{a \link{character} string}

\item{membership}{a \link{vector} of length \code{nHabitats}}

\item{searchQ}{a \link{vector} of length \code{nHabitats}}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_L} for the trace model
}
