% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{setup_Linits.basic}
\alias{setup_Linits.basic}
\title{Setup the basic model}
\usage{
\method{setup_Linits}{basic}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_Linits} for the basic model
}
