% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{setup_Linits.trace}
\alias{setup_Linits.trace}
\title{Setup the trace model}
\usage{
\method{setup_Linits}{trace}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_Linits} for the trace model
}
