% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fraction.R
\name{setup_local_fraction_simple}
\alias{setup_local_fraction_simple}
\title{Set up the local_fraction for static models}
\usage{
setup_local_fraction_simple(pars, local_frac = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{local_frac}{is the fraction of the ambient human / host population that is not a visitor}
}
\value{
\link{list}
}
\description{
The local fraction
}
