% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{setup_Lpar.basic}
\alias{setup_Lpar.basic}
\title{Setup Lpar for the basic model}
\usage{
\method{setup_Lpar}{basic}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Lpar} for the basic model
}
