% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{setup_Lpar}
\alias{setup_Lpar}
\title{A function to set up adult mosquito models}
\usage{
setup_Lpar(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{Lname}.
}
