% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-null.R
\name{setup_lsm_null}
\alias{setup_lsm_null}
\title{Make parameters for the null model of LSM (do nothing)}
\usage{
setup_lsm_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Make parameters for the null model of LSM (do nothing)
}
