% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R, R/import-null.R
\name{setup_MalariaImportation.null}
\alias{setup_MalariaImportation.null}
\title{A function to set up malaria importation}
\usage{
\method{setup_MalariaImportation}{null}(pars, IMname, IMopts = list())

\method{setup_MalariaImportation}{null}(pars, IMname, IMopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{IMname}{a \link{character} string}

\item{IMopts}{a \link{list}}
}
\value{
a \link{list}

a \link{list}
}
\description{
Implements \link{setup_MalariaImportation} for the basic model

Implements \link{setup_MalariaImportation} for the null model
}
