% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-interface.R
\name{setup_MalariaImportation}
\alias{setup_MalariaImportation}
\title{A function to set up malaria importation}
\usage{
setup_MalariaImportation(pars, IMname, IMopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{IMname}{a \link{character} string}

\item{IMopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on \code{IMname}.
}
