% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_medical.R
\name{setup_mass_medical_forced}
\alias{setup_mass_medical_forced}
\title{Set up a model for mass medical}
\usage{
setup_mass_medical_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up a model for mass medical
}
