% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_medical.R
\name{setup_mass_medical_null}
\alias{setup_mass_medical_null}
\title{Set up the null model for mass medical (do nothing)}
\usage{
setup_mass_medical_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for mass medical (do nothing)
}
