% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{setup_MYZ.GeRM}
\alias{setup_MYZ.GeRM}
\title{Setup the GeRM model for adult mosquitoes}
\usage{
\method{setup_MYZ}{GeRM}(pars, MYZname, nPatches = 1, MYZopts = list(), calK = diag(1))
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZname}{a \link{character} string}

\item{nPatches}{a \link{numeric} value (an integer), the number of patches}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZ} for the GeRM model
}
