% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{setup_MYZ.Gtrace}
\alias{setup_MYZ.Gtrace}
\title{Setup the Gtrace}
\usage{
\method{setup_MYZ}{Gtrace}(pars, MYZname, nPatches = 1, MYZopts = NULL, calK = diag(1))
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZname}{a \link{character} string}

\item{nPatches}{a \link{numeric} value (an integer), the number of patches}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZ} for the Gtrace model
}
