% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{setup_MYZ}
\alias{setup_MYZ}
\title{A function to set up adult mosquito models}
\usage{
setup_MYZ(pars, MYZname, nPatches = 1, MYZopts = list(), calK = diag(1))
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZname}{a \link{character} string}

\item{nPatches}{a \link{numeric} value (an integer), the number of patches}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYZname}.
}
