% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{setup_MYZ.RM}
\alias{setup_MYZ.RM}
\title{Setup the RM model}
\usage{
\method{setup_MYZ}{RM}(pars, MYZname, nPatches = 1, MYZopts = list(), calK = diag(1))
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZname}{a \link{character} string}

\item{nPatches}{a \link{numeric} value (an integer), the number of patches}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZ} for the RM model
}
