% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{setup_MYZinits.basicM}
\alias{setup_MYZinits.basicM}
\title{Setup the basicM model}
\usage{
\method{setup_MYZinits}{basicM}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the basicM model
}
