% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{setup_MYZinits.GeRM}
\alias{setup_MYZinits.GeRM}
\title{Setup the GeRM model}
\usage{
\method{setup_MYZinits}{GeRM}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the GeRM model
}
