% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{setup_MYZinits.Gtrace}
\alias{setup_MYZinits.Gtrace}
\title{Setup the Gtrace model}
\usage{
\method{setup_MYZinits}{Gtrace}(pars, s, MYZopts = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the Gtrace model
}
