% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{setup_MYZinits.Ztrace}
\alias{setup_MYZinits.Ztrace}
\title{Setup the Ztrace model}
\usage{
\method{setup_MYZinits}{Ztrace}(pars, s, MYZopts = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the Ztrace model
}
