% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{setup_MYZpar.GeRM}
\alias{setup_MYZpar.GeRM}
\title{Setup MYZpar for the GeRM model}
\usage{
\method{setup_MYZpar}{GeRM}(MYZname, pars, s, MYZopts = list(), EIPmod, calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}

\item{EIPmod}{is a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the GeRM model
}
