% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{setup_MYZpar}
\alias{setup_MYZpar}
\title{A function to set up adult mosquito models}
\usage{
setup_MYZpar(
  MYZname,
  pars,
  s,
  MYZopts = list(),
  EIPmod = setup_eip_static(),
  calK = diag(1)
)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}

\item{EIPmod}{is a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYZname}.
}
