% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{setup_MYZpar.RM}
\alias{setup_MYZpar.RM}
\title{Setup MYZpar for the RM model}
\usage{
\method{setup_MYZpar}{RM}(MYZname, pars, s, MYZopts = list(), EIPmod, calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}

\item{EIPmod}{is a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the RM model
}
