% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{setup_other_blood_hosts_null}
\alias{setup_other_blood_hosts_null}
\title{Make parameters for the null model for other blood hosts (do nothing)}
\usage{
setup_other_blood_hosts_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the null model for other blood hosts (do nothing)
}
