% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{setup_other_blood_hosts_static}
\alias{setup_other_blood_hosts_static}
\title{Make parameters for the static model for other blood hosts (do nothing)}
\usage{
setup_other_blood_hosts_static(pars, Other = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{Other}{the availability of other blood hosts}
}
\value{
\link{list}
}
\description{
Make parameters for the static model for other blood hosts (do nothing)
}
