% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_blood_hosts.R
\name{setup_otherblood_forced}
\alias{setup_otherblood_forced}
\title{Set up forcing for the availability of other blood hosts}
\usage{
setup_otherblood_forced(
  pars,
  otherblood_Opts = list(),
  otherblood_scale = 0,
  otherblood_t = NULL
)
}
\arguments{
\item{pars}{a \link{list}}

\item{otherblood_Opts}{a list, overwrites default values}

\item{otherblood_scale}{a non-negative numeric value to scale the mean availability of other blood hosts}

\item{otherblood_t}{a function that specifies the temporal pattern for the availability of other blood hosts}
}
\value{
none
}
\description{
Set up forcing for the availability of other blood hosts
}
