% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovitrap.R
\name{setup_ovitraps_null}
\alias{setup_ovitraps_null}
\title{Set up the null model for oviposition traps (do nothing)}
\usage{
setup_ovitraps_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for oviposition traps (do nothing)
}
