% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_rainfall_null}
\alias{setup_rainfall_null}
\title{Set up the null model for RAINFALL}
\usage{
setup_rainfall_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for RAINFALL
}
