% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_relative_humidity_null}
\alias{setup_relative_humidity_null}
\title{Set up the null model for HUMIDITY}
\usage{
setup_relative_humidity_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for HUMIDITY
}
