% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-forced.R
\name{setup_resources_forced}
\alias{setup_resources_forced}
\title{Set up a model for mass medical}
\usage{
setup_resources_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up a model for mass medical
}
