% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-null.R
\name{setup_resources_null}
\alias{setup_resources_null}
\title{Set up parameters for the null model for resource availability}
\usage{
setup_resources_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Set up parameters for the null model for resource availability
}
