% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{setup_shock_null}
\alias{setup_shock_null}
\title{Set up the null model for shocks (do nothing)}
\usage{
setup_shock_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for shocks (do nothing)
}
