% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits-interface.R
\name{setup_sugar_baits_null}
\alias{setup_sugar_baits_null}
\title{Set up the null model for sugar baits (do nothing)}
\usage{
setup_sugar_baits_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for sugar baits (do nothing)
}
