% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{setup_sugar_forced}
\alias{setup_sugar_forced}
\title{Set up forcing for the availability of sugar}
\usage{
setup_sugar_forced(pars, sugar_Opts = list(), sugar_scale = 0, sugar_t = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{sugar_Opts}{a list, overwrites default values}

\item{sugar_scale}{a non-negative numeric value to scale the mean availability of sugar}

\item{sugar_t}{a function that specifies the temporal pattern for the availability of sugar}
}
\value{
none
}
\description{
Set up forcing for the availability of sugar
}
