% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{setup_sugar_null}
\alias{setup_sugar_null}
\title{Make parameters for the null model for sugar (do nothing)}
\usage{
setup_sugar_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the null model for sugar (do nothing)
}
