% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_dynamics.R
\name{setup_sugar_static}
\alias{setup_sugar_static}
\title{Make parameters for the static model for sugar (do nothing)}
\usage{
setup_sugar_static(pars, Sugar = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{Sugar}{describes sugar availability}
}
\value{
\link{list}
}
\description{
Make parameters for the static model for sugar (do nothing)
}
