% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_temperature_null}
\alias{setup_temperature_null}
\title{Set up the null model for temperature}
\usage{
setup_temperature_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for temperature
}
