% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{setup_Travel_basic}
\alias{setup_Travel_basic}
\title{Make parameters for the basic model for Travel Malaria}
\usage{
setup_Travel_basic(pars, TravelOpts = list(), delta = 0, delta_t = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{TravelOpts}{a list, overwrites default values}

\item{delta}{the mean FoI for travel malaria}

\item{delta_t}{the temporal signal (should have a mean of 1)}
}
\value{
none
}
\description{
Make parameters for the basic model for Travel Malaria
}
