% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-eir.R
\name{setup_travel_eir}
\alias{setup_travel_eir}
\title{Set up parameters and function for an eir-based travel_malaria model}
\usage{
setup_travel_eir(
  pars,
  travel_Opts = list(),
  frac_time_spent_traveling = 0.01,
  travel_eir_scale = 0,
  travel_eir_t = NULL
)
}
\arguments{
\item{pars}{a \link{list}}

\item{travel_Opts}{a list, overwrites default values}

\item{frac_time_spent_traveling}{is the fraction of time spent travel}

\item{travel_eir_scale}{a non-negative numeric value to scale the mean eir experienced while traveling}

\item{travel_eir_t}{the temporal pattern for eir experienced while traveling}
}
\value{
none
}
\description{
Set up parameters and function for an eir-based travel_malaria model
}
