% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-foi.R
\name{setup_travel_foi}
\alias{setup_travel_foi}
\title{Set up parameters and function for the FoI-based travel_malaria model}
\usage{
setup_travel_foi(pars, travel_Opts = list(), delta_scale = 0, delta_t = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{travel_Opts}{a list, overwrites default values}

\item{delta_scale}{a non-negative numeric value to scale the mean travel_malaria FoI}

\item{delta_t}{the temporal pattern for travel_malaria FoI}
}
\value{
none
}
\description{
Set up parameters and function for the FoI-based travel_malaria model
}
