% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-null.R
\name{setup_travel_null}
\alias{setup_travel_null}
\title{A function to set up malaria importation}
\usage{
setup_travel_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Setup a null model for travel malaria
}
