% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorcontrol-control.R
\name{setup_vc_control}
\alias{setup_vc_control}
\title{Make parameters for the control model of vector control (do nothing)}
\usage{
setup_vc_control(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the control model of vector control (do nothing)
}
