% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-null.R
\name{setup_vc_null}
\alias{setup_vc_null}
\title{Distribute vector control, the null model}
\usage{
setup_vc_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Distribute vector control, the null model
}
