% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{setup_Visitors_basic}
\alias{setup_Visitors_basic}
\title{Make parameters for the basic model for visitors}
\usage{
setup_Visitors_basic(
  pars,
  IMopts,
  local_frac = 1,
  xv = 0,
  xt = NULL,
  Vm = 0,
  Vt = NULL
)
}
\arguments{
\item{pars}{a \link{list}}

\item{IMopts}{a \link{list}}

\item{local_frac}{the fraction of bites taken on residents}

\item{xv}{the mean probability a bite on a visitor infects a mosquito}

\item{xt}{a function to force x_visitors over time}

\item{Vm}{is the mean availability of visitors for blood feeding}

\item{Vt}{a function to force availability of Visitors over time}
}
\value{
none
}
\description{
Make parameters for the basic model for visitors
}
