% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors_static}
\alias{setup_visitors_static}
\title{Make parameters for the static model visitors (no visitors)}
\usage{
setup_visitors_static(pars, local_frac = 1, Visitors = 0, x_visitors = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{local_frac}{is the fraction of humans / hosts that are not visitors}

\item{Visitors}{is the availability of visitors}

\item{x_visitors}{is the net infectiousness of the visitors}
}
\value{
\link{list}
}
\description{
Make parameters for the static model visitors (no visitors)
}
