% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_weather_forced}
\alias{setup_weather_forced}
\title{Set up the forced model for weather}
\usage{
setup_weather_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the forced model for weather
}
