% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{setup_X.hMoI}
\alias{setup_X.hMoI}
\title{Setup Xpar.hMoI}
\usage{
\method{setup_X}{hMoI}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_X} for the hMoI model
}
