% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{setup_X.SIP}
\alias{setup_X.SIP}
\title{Setup Xpar.SIP}
\usage{
\method{setup_X}{SIP}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_X} for the SIP model
}
