% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{setup_Xinits}
\alias{setup_Xinits}
\title{A function to set up Xpar}
\usage{
setup_Xinits(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on \code{Xname}.
}
