% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{setup_Xpar.hMoI}
\alias{setup_Xpar.hMoI}
\title{Setup Xpar.hMoI}
\usage{
\method{setup_Xpar}{hMoI}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the hMoI model
}
