% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{setup_Xpar.SIS}
\alias{setup_Xpar.SIS}
\title{Setup Xpar.SIS}
\usage{
\method{setup_Xpar}{SIS}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the SIS model
}
