% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{setup_Xpar.trace}
\alias{setup_Xpar.trace}
\title{Setup Xpar.trace}
\usage{
\method{setup_Xpar}{trace}(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the SIS model
}
