% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shapeIt}
\alias{shapeIt}
\title{Check the shape and dimensions of an object}
\usage{
shapeIt(obj, d1, d2)
}
\arguments{
\item{obj}{a \link{numeric} object}

\item{d1}{an \link{integer}}

\item{d2}{an \link{integer}}
}
\value{
\link{matrix}
}
\description{
Check the shape and dimensions of an object
}
