% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{Shock}
\alias{Shock}
\title{Set up shocks}
\usage{
Shock(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$SHOCK}.
}
