% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-null.R
\name{SprayHouses.null}
\alias{SprayHouses.null}
\title{Do mass house spraying (IRS)}
\usage{
\method{SprayHouses}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$IRS}.
}
