% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-interface.R
\name{SprayHouses}
\alias{SprayHouses}
\title{Do mass house spraying (IRS)}
\usage{
SprayHouses(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$IRS}.
}
