% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{Sugar.null}
\alias{Sugar.null}
\title{Set the values of exogenous variables describing sugar}
\usage{
\method{Sugar}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Sugar} for the null model of sugar (do nothing)
}
