% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{Sugar}
\alias{Sugar}
\title{Set the values of exogenous variables describing sugar}
\usage{
Sugar(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$SUGAR}.
}
