% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-forced.R
\name{SugarAvailability.forced}
\alias{SugarAvailability.forced}
\title{Compute total availability of sugar}
\usage{
\method{SugarAvailability}{forced}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of sugar
}
