% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-null.R
\name{SugarAvailability.null}
\alias{SugarAvailability.null}
\title{Compute total availability of sugar}
\usage{
\method{SugarAvailability}{null}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Computes the availability of sugar for the null model (do nothing)
}
