% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{SugarAvailability}
\alias{SugarAvailability}
\title{Update the availability of sugar}
\usage{
SugarAvailability(pars)
}
\arguments{
\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$RESOURCES}.
}
