% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits-null.R
\name{SugarBaitEffects.null}
\alias{SugarBaitEffects.null}
\title{Methods for the effects of the sugar baits}
\usage{
\method{SugarBaitEffects}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{SugarBaitEffects} for the null model (do nothing)
}
