% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits-interface.R
\name{SugarBaitEffects}
\alias{SugarBaitEffects}
\title{Methods for the durability and effects of the sugar baits}
\usage{
SugarBaitEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$SUGAR_BAITS}.
}
