% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits-null.R
\name{SugarBaitEffectSizes.null}
\alias{SugarBaitEffectSizes.null}
\title{Methods for the effect sizes of the sugar baits}
\usage{
\method{SugarBaitEffectSizes}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$SUGAR_BAITS}.
}
