% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits-null.R
\name{SugarBaits.null}
\alias{SugarBaits.null}
\title{Methods for sugar baits}
\usage{
\method{SugarBaits}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{SugarBaits} for the null model (do nothing)
}
