% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_dynamics.R
\name{SugarDynamics.static}
\alias{SugarDynamics.static}
\title{Set the values of exogenous variables describing sugar}
\usage{
\method{SugarDynamics}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{SugarDynamics} for the static model of sugar (do nothing)
}
