% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Temperature.null}
\alias{Temperature.null}
\title{Methods for exogenous variables describing temperature}
\usage{
\method{Temperature}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Temperature} for the null model (no variables)
}
