% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission}
\alias{Transmission}
\title{Compute the biting distribution matrix, beta}
\usage{
Transmission(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars, }{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}
}
