% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission.static}
\alias{Transmission.static}
\title{Compute transmission terms dynamically, no update required}
\usage{
\method{Transmission}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$BFpar}
}
