% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-basic.R
\name{travel_foi.basic}
\alias{travel_foi.basic}
\title{travel_foi, a null model}
\usage{
\method{travel_foi}{basic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_foi} for the null model (do nothing)
}
