% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-null.R
\name{travel_foi.null}
\alias{travel_foi.null}
\title{travel_foi, a null model}
\usage{
\method{travel_foi}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_foi} for the null model (do nothing)
}
