% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-interface.R
\name{travel_foi}
\alias{travel_foi}
\title{travel_foi}
\usage{
travel_foi(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} value
}
\description{
This method dispatches on the type of \code{pars$Ipar}.
}
