% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-eir.R
\name{travel_malaria.eir}
\alias{travel_malaria.eir}
\title{A model for travel malaria based on the eir in destinations}
\usage{
\method{travel_malaria}{eir}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_malaria} through a model for an eir-based model
}
