% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-foi.R
\name{travel_malaria.foi}
\alias{travel_malaria.foi}
\title{A model for the travel FoI}
\usage{
\method{travel_malaria}{foi}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_malaria} through a model for the travel FoI
}
