% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-null.R
\name{travel_malaria.null}
\alias{travel_malaria.null}
\title{travel, a null model}
\usage{
\method{travel_malaria}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_malaria} for the null model (do nothing)
}
