% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-interface.R
\name{travel_malaria}
\alias{travel_malaria}
\title{Simulate travel malaria}
\usage{
travel_malaria(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
the travel FoI, a \link{numeric} value
}
\description{
This method dispatches on the type of \code{pars$TRAVEL}.
}
