% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-static.R
\name{travel_malaria.static}
\alias{travel_malaria.static}
\title{travel, a static model}
\usage{
\method{travel_malaria}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_malaria} for the static model (do nothing)
}
