% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-null.R
\name{TreatHabitats.null}
\alias{TreatHabitats.null}
\title{Set up the null model for LSM}
\usage{
\method{TreatHabitats}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSM}
}
