% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-interface.R
\name{TreatHabitats}
\alias{TreatHabitats}
\title{Modify values due to treat habitats as part of LSM,
called by Control->VectorControl}
\usage{
TreatHabitats(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$LSM}.
}
