% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_BloodFeeding.dynamic}
\alias{update_BloodFeeding.dynamic}
\title{Update total availability of blood hosts for dynamical models}
\usage{
update_BloodFeeding.dynamic(t, y, pars)
}
\value{
a \link{list}
}
\description{
Implements \link{update_BloodFeeding} for a dynamic model
}
