% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-static.R
\name{update_BloodFeeding.static}
\alias{update_BloodFeeding.static}
\title{Update blood feeding for static models}
\usage{
\method{update_BloodFeeding}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_BloodFeeding} for a dynamic model
}
