% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{update_EggLaying}
\alias{update_EggLaying}
\title{Update the availability of habitats and egg laying}
\usage{
update_EggLaying(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Q}. It
computes availability of all blood hosts at time \code{t}
}
