% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_inits}
\alias{update_inits}
\title{Set the initial values to the last values of the last simulation}
\usage{
update_inits(y0, pars)
}
\arguments{
\item{y0}{a \link{vector} of initial values}

\item{pars}{a \link{list}}
}
\value{
y a \link{numeric} vector
}
\description{
Set the initial values to the last values of the last simulation
}
