% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{update_inits_H.dynamic}
\alias{update_inits_H.dynamic}
\title{Update inits for the static human demography model}
\usage{
\method{update_inits_H}{dynamic}(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
Update inits for the static human demography model
}
