% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{update_inits_H}
\alias{update_inits_H}
\title{Set the initial values as a vector}
\usage{
update_inits_H(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Hpar}.
}
