% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{update_inits_H.static}
\alias{update_inits_H.static}
\title{Update inits for the static human demography model}
\usage{
\method{update_inits_H}{static}(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
Update inits for the static human demography model
}
