% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{update_inits_MYZ.GeRM_dde}
\alias{update_inits_MYZ.GeRM_dde}
\title{Update the initial values for GeRM adult mosquito model}
\usage{
\method{update_inits_MYZ}{GeRM_dde}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update the initial values for GeRM adult mosquito model
}
