% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{update_inits_MYZ.Gtrace}
\alias{update_inits_MYZ.Gtrace}
\title{Update inits for Gtrace}
\usage{
\method{update_inits_MYZ}{Gtrace}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for Gtrace
}
