% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{update_inits_MYZ.Ztrace}
\alias{update_inits_MYZ.Ztrace}
\title{Update inits for Ztrace}
\usage{
\method{update_inits_MYZ}{Ztrace}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for Ztrace
}
