% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{update_inits_X.hMoI}
\alias{update_inits_X.hMoI}
\title{Update inits for hybrid MoI human model from a vector of states}
\usage{
\method{update_inits_X}{hMoI}(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
Update inits for hybrid MoI human model from a vector of states
}
