% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{update_inits_X}
\alias{update_inits_X}
\title{Set the initial values from a vector of states}
\usage{
update_inits_X(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
