% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_O.dynamic}
\alias{update_O.dynamic}
\title{Update other host availability dynamically}
\usage{
\method{update_O}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_O} for a static model
}
