% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{update_O}
\alias{update_O}
\title{Update the availability of other blood hosts}
\usage{
update_O(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$O}. It
computes availability of blood hosts at time \code{t}. Other blood hosts
include any vertebrate host that does not serve as a host for the pathogen.
}
