% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-static.R
\name{update_O.static}
\alias{update_O.static}
\title{Update other host availability for static models}
\usage{
\method{update_O}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_O} when pars$O doesn't change
}
