% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_S.dynamic}
\alias{update_S.dynamic}
\title{Dynamically update the total availability of sugar}
\usage{
update_S.dynamic(t, pars)
}
\value{
a \link{list}
}
\description{
Implements \link{update_S} for a dynamic model
}
