% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{update_S}
\alias{update_S}
\title{Update the total availability of sugar}
\usage{
update_S(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$S}. It should
update, and if necessary compute, the availability of sugar at time \code{t}
}
