% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-static.R
\name{update_S.static}
\alias{update_S.static}
\title{Update sugar availability for static models}
\usage{
\method{update_S}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_S} when pars$S doesn't change
}
