% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_W.dynamic}
\alias{update_W.dynamic}
\title{Update host availability dynamically}
\usage{
\method{update_W}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_W} for a static model
}
