% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-interface.R
\name{update_W}
\alias{update_W}
\title{Update host availability}
\usage{
update_W(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$W}. It should
update the availability of the pathogen's hosts for blood feeding at time \code{t}
}
