% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-static.R
\name{update_W.static}
\alias{update_W.static}
\title{Update host availability for static models}
\usage{
\method{update_W}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_W} when pars$W doesn't change
}
