% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc-lemenach.R
\name{VectorControl.lemenach}
\alias{VectorControl.lemenach}
\title{Modify baseline values due to vector control}
\usage{
\method{VectorControl}{lemenach}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControl} for the Le Menach ITN model of vector control
}
