% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc-null.R
\name{VectorControl.null}
\alias{VectorControl.null}
\title{Modify baseline values due to vector control}
\usage{
\method{VectorControl}{null}(t, y, pars, MosyBehavior)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{MosyBehavior}{values returned by \link{MosquitoBehavior}}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControl} for the null model of vector control (do nothing)
}
