% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc-interface.R
\name{VectorControl}
\alias{VectorControl}
\title{Modify baseline values due to vector control}
\usage{
VectorControl(t, y, pars, MosyBehavior)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{MosyBehavior}{values returned by \link{MosquitoBehavior}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$VCpar}. It takes the
baseline \code{MosyBehavior} values and modifies them, potentially at multiple time
points for models with delay.
}
