% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-control.R
\name{VectorControlEffects.control}
\alias{VectorControlEffects.control}
\title{Vector control durability and effects}
\usage{
\method{VectorControlEffects}{control}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControlEffects} for the control model of vector control (do nothing)
}
