% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-interface.R
\name{VectorControlEffects}
\alias{VectorControlEffects}
\title{Vector control: durability & effects}
\usage{
VectorControlEffects(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$VECTOR_CONTROL}.
}
