% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-control.R
\name{VectorControlEffectSizes.control}
\alias{VectorControlEffectSizes.control}
\title{Distribute vector control, the null model}
\usage{
\method{VectorControlEffectSizes}{control}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControlEffectSizes} for the control model of vector control (do nothing)
}
