% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-null.R
\name{VectorControlEffectSizes.null}
\alias{VectorControlEffectSizes.null}
\title{Vector control effect sizes}
\usage{
\method{VectorControlEffectSizes}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControlEffectSizes} for the null model of vector control (do nothing)
}
