% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors.null}
\alias{Visitors.null}
\title{Visitors, a null model}
\usage{
\method{Visitors}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{Visitors} for the null model (do nothing)
}
