% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors}
\alias{Visitors}
\title{Visitors}
\usage{
Visitors(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$VISITORS}.
}
