% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors.static}
\alias{Visitors.static}
\title{Visitors, a static model}
\usage{
\method{Visitors}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{Visitors} for the static model (do nothing)
}
