% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-null.R
\name{Weather.null}
\alias{Weather.null}
\title{Modify parameters due to exogenous forcing}
\usage{
\method{Weather}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{Weather} for the null model of weather (do nothing)
}
