% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-interface.R
\name{Weather}
\alias{Weather}
\title{Set the values of exogenous variables describing weather}
\usage{
Weather(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$WETpar}.
}
