% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqn.R
\name{xDE_diffeqn_aquatic}
\alias{xDE_diffeqn_aquatic}
\title{Differential equation models for aquatic mosquito populations}
\usage{
xDE_diffeqn_aquatic(t, y, pars, F_eta)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{F_eta}{trace function that returns eggs laid}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
