% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqn.R
\name{xDE_diffeqn_cohort}
\alias{xDE_diffeqn_cohort}
\title{Differential equation models for human cohorts}
\usage{
xDE_diffeqn_cohort(a, y, pars, F_eir)
}
\arguments{
\item{a}{age of a cohort}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{F_eir}{a trace function that returns the eir}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
