% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{xde_lines_aEIR}
\alias{xde_lines_aEIR}
\title{Add lines for the annualized EIR \emph{vs.} time}
\usage{
xde_lines_aEIR(terms, pars, clrs = "black", llty = 1)
}
\arguments{
\item{terms}{a list with the outputs of \code{parse_deout()}}

\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{exDE::xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the annualized EIR \emph{vs.} time
}
