% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{xde_lines_EIR}
\alias{xde_lines_EIR}
\title{Add lines for the EIR \emph{vs.} time}
\usage{
xde_lines_EIR(tm, EIR, nStrata, clrs = "black", llty = 1)
}
\arguments{
\item{tm}{the time}

\item{EIR}{the entomological inoculation rate}

\item{nStrata}{the number of human / host population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the EIR \emph{vs.} time
}
