% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{xde_lines_M}
\alias{xde_lines_M}
\title{Add lines for adult mosquito population density}
\usage{
xde_lines_M(MYZ, pars, clrs = "darkblue", llty = 1)
}
\arguments{
\item{MYZ}{a \link{list} the output of \code{exDE::parse_deout()}}

\item{pars}{a \link{list} that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for adult mosquito population density
}
