% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{xde_lines_PR}
\alias{xde_lines_PR}
\title{Add lines for the prevalence / parasite rate (PR) from a model of human infection and immunity}
\usage{
xde_lines_PR(terms, pars, clrs = "black", llty = 1)
}
\arguments{
\item{terms}{a list with the outputs of \code{exDE::parse_deout()}}

\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) that specifies \code{lty} for plotting}
}
\description{
Add lines for the prevalence / parasite rate (PR) from a model of human infection and immunity
}
