% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{xde_lines_X.SIS}
\alias{xde_lines_X.SIS}
\title{Add lines for the density of infected individuals for the SIS model}
\usage{
\method{xde_lines_X}{SIS}(XH, pars, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{XH}{a list with the outputs of \code{exDE::parse_deout()}}

\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIS model
}
