% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{xde_lines_YZ_fracs}
\alias{xde_lines_YZ_fracs}
\title{Add lines for the fraction of infected and infective mosquitoes}
\usage{
xde_lines_YZ_fracs(MYZ, pars, Yclrs = "purple", Zclrs = "darkred", llty = 1)
}
\arguments{
\item{MYZ}{a \link{list} the output of \code{exDE::parse_deout()}}

\item{pars}{a \link{list} specifying the model}

\item{Yclrs}{a \link{list} of colors}

\item{Zclrs}{a \link{list} of colors}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the fraction of infected and infective mosquitoes
}
