% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{xde_lines_Z_fracs}
\alias{xde_lines_Z_fracs}
\title{Add lines for the fraction of infected and infective mosquitoes}
\usage{
xde_lines_Z_fracs(MZ, pars, Zclrs = "darkred", llty = 1)
}
\arguments{
\item{MZ}{a list the output of \code{exDE::parse_deout()}}

\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{Zclrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the fraction of infected and infective mosquitoes
}
