% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{xde_plot_aEIR}
\alias{xde_plot_aEIR}
\title{Plot the annualized EIR \emph{vs.} time}
\usage{
xde_plot_aEIR(pars, clrs = "black", llty = 1, stable = FALSE, add_axes = TRUE)
}
\arguments{
\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{exDE::xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the annualized EIR \emph{vs.} time
}
