% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{xde_plot_M}
\alias{xde_plot_M}
\title{Plot adult mosquito population density}
\usage{
xde_plot_M(pars, clrs = "darkblue", llty = 1, stable = FALSE, add_axes = TRUE)
}
\arguments{
\item{pars}{a \link{list} that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}

\item{stable}{a \link{logical} set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a \link{logical} to plot axes only if TRUE}
}
\description{
Plot adult mosquito population density
}
