% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{xde_plot_X.SIP}
\alias{xde_plot_X.SIP}
\title{Plot the density of infected individuals for the SIP model}
\usage{
\method{xde_plot_X}{SIP}(
  pars,
  clrs = c("black", "darkgreen"),
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a \link{list} that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a \link{character} vector of colors}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}

\item{stable}{a \link{logical} set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a \link{logical} to plot axes only if TRUE}
}
\description{
Plot the density of infected individuals for the SIP model
}
