% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{xde_plot_X.SIS}
\alias{xde_plot_X.SIS}
\title{Plot the density of infected individuals for the SIS model}
\usage{
\method{xde_plot_X}{SIS}(
  pars,
  i,
  clrs = c("darkblue", "darkred"),
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the density of infected individuals for the SIS model
}
