% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{xde_plot_YZ_fracs}
\alias{xde_plot_YZ_fracs}
\title{Plot the fraction of infected and infective mosquitoes}
\usage{
xde_plot_YZ_fracs(
  pars,
  Yclrs = "purple",
  Zclrs = "darkred",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{Yclrs}{a vector of colors for infected mosquitoes}

\item{Zclrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical, set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical to plot axes only if TRUE}
}
\description{
Plot the fraction of infected and infective mosquitoes
}
