% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xde_setup}
\alias{xde_setup}
\title{Set up a model for xde_diffeqn}
\usage{
xde_setup(
  modelName,
  MYZname = "RM",
  Xname = "SIS",
  Lname = "trace",
  nPatches = 1,
  nVectors = 1,
  nHosts = 1,
  HPop = 1000,
  membership = 1,
  MYZopts = list(),
  calK = "herethere",
  calKopts = list(),
  EIPname = "static",
  EIPopts = list(),
  Xopts = list(),
  Hopts = list(),
  residence = 1,
  searchB = 1,
  TaR = "athome",
  TaRopts = list(),
  searchQ = 1,
  Lopts = list()
)
}
\arguments{
\item{modelName}{is a name for the model (arbitrary)}

\item{MYZname}{is a character string defining a MYZ model}

\item{Xname}{is a character string defining a X model}

\item{Lname}{is a character string defining a L model}

\item{nPatches}{is the number of patches}

\item{nVectors}{is the number of vector species}

\item{nHosts}{is the number of vertebrate host species}

\item{HPop}{is the number of humans in each patch}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{MYZopts}{a list to configure the MYZ model}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{calKopts}{are the options to setup calK}

\item{EIPname}{are the options to setup EIPmod}

\item{EIPopts}{are the options to setup EIPmod}

\item{Xopts}{a list to configure the X model}

\item{Hopts}{a list to configure the H model}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{searchB}{is a vector of search weights for blood feeding}

\item{TaR}{is either a TaR matrix or a string to call a function that sets it up}

\item{TaRopts}{are the options to setup TaR}

\item{searchQ}{is a vector of search weights for egg laying}

\item{Lopts}{a list to configure the L model}
}
\value{
a \link{list}
}
\description{
Set up a model for xde_diffeqn
}
