% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xde_setup_aquatic}
\alias{xde_setup_aquatic}
\title{Set up a model for xde_diffeqn_aqua}
\usage{
xde_setup_aquatic(
  modelName,
  nHabitats = 1,
  nVectors = 1,
  Lname = "basic",
  Lopts = list(),
  MYZopts = list(),
  LSMname = "null"
)
}
\arguments{
\item{modelName}{is a name for the model (arbitrary)}

\item{nHabitats}{is the number of habitats}

\item{nVectors}{is the number of vector species}

\item{Lname}{is a character string defining a L model}

\item{Lopts}{a list to configure the L model}

\item{MYZopts}{a list to configure F_eggs from the Gtrace model}

\item{LSMname}{is a character string defining a LSM model}
}
\value{
a \link{list}
}
\description{
Set up a model for xde_diffeqn_aqua
}
