% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xde_setup_cohort}
\alias{xde_setup_cohort}
\title{Set up a model for xde_diffeqn_cohort}
\usage{
xde_setup_cohort(
  modelName,
  F_eir,
  Xname = "SIS",
  HPop = 1000,
  Xopts = list(),
  Hopts = list(),
  eir_out = TRUE,
  NI_out = TRUE
)
}
\arguments{
\item{modelName}{is a name for the model (arbitrary)}

\item{F_eir}{is a function F_eir(t, pars) that returns the daily FoI}

\item{Xname}{is a character string defining a X model}

\item{HPop}{is the number of humans in each patch}

\item{Xopts}{a list to configure the X model}

\item{Hopts}{a list to configure the H model}

\item{eir_out}{the cumulative eir is output if TRUE}

\item{NI_out}{the cumulative NI is output if TRUE}
}
\value{
a \link{list}
}
\description{
Set up a model for xde_diffeqn_cohort
}
