% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xde_setup_mosy}
\alias{xde_setup_mosy}
\title{Set up a model for xde_diffeqn_mosy}
\usage{
xde_setup_mosy(
  modelName = "unnamed",
  MYZname = "basicM",
  Lname = "basic",
  nPatches = 1,
  nVectors = 1,
  membership = 1,
  MYZopts = list(),
  calK = "herethere",
  calKopts = list(),
  searchQ = 1,
  Lopts = list(),
  kappa = NULL
)
}
\arguments{
\item{modelName}{is a name for the model (arbitrary)}

\item{MYZname}{is a character string defining a MYZ model}

\item{Lname}{is a character string defining a L model}

\item{nPatches}{is the number of patches}

\item{nVectors}{is the number of vector species}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{MYZopts}{a list to configure the MYZ model}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{calKopts}{are the options to setup calK}

\item{searchQ}{is a vector of search weights for egg laying}

\item{Lopts}{a list to configure the L model}

\item{kappa}{values -- net infectivity to force adult infection dynamics}
}
\value{
a \link{list}
}
\description{
Set up a model for xde_diffeqn_mosy
}
