% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{xde_solve.aqua_dde}
\alias{xde_solve.aqua_dde}
\title{Solve a system of delay equations for aquatic dynamics, forced by egg deposition, using xDE_diffeqn_aquatic}
\usage{
\method{xde_solve}{aqua_dde}(pars, Tmax = 365, dt = 1)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}
}
\value{
a \link{matrix}
}
\description{
Implements \link{xde_solve} for aquatic dynaamic
}
