% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{xde_solve.ode}
\alias{xde_solve.ode}
\title{Solve a system of equations as an ode}
\usage{
\method{xde_solve}{ode}(pars, Tmax = 365, dt = 1)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}
}
\value{
a \link{list}
}
\description{
Implements \link{xde_solve} for ordinary differential equations
}
