% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{xde_solve}
\alias{xde_solve}
\title{Solve a system of equations}
\usage{
xde_solve(pars, Tmax = 365, dt = 1)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}.
}
