% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{xde_stable_orbit}
\alias{xde_stable_orbit}
\title{Solve for the steady state or stable orbit of a system of equations}
\usage{
xde_stable_orbit(pars, Ymax = 10)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Ymax}{the number of years to burn-in}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}.
}
