% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solving.R
\name{xde_steady}
\alias{xde_steady}
\title{Solve for the steady state of a system of equations using \link[rootSolve:steady]{rootSolve::steady}}
\usage{
xde_steady(pars)
}
\arguments{
\item{pars}{a \link{list} that defines a model}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}
}
