% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{Bionomics}
\alias{Bionomics}
\title{Set bionomic parameter rates relative to baseline}
\usage{
Bionomics(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This calls Mbionomics and Lbionmics for each species
}
