% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-basic.R
\name{Births.basic}
\alias{Births.basic}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{Births}{basic}(t, y, pars)
}
\value{
a \link{numeric} vector of length \code{nStrata} or of length 0
}
\description{
Implements \link{dHdt} for the forced (basic) model.
}
