% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{Births}
\alias{Births}
\title{A function that computes the birth rate for human populations}
\usage{
Births(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of \code{y}.
}
