% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_beta}
\alias{compute_beta}
\title{Compute beta}
\usage{
compute_beta(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars, }{a \link{list}}
}
\value{
\link{list}
}
\description{
This function computes the mixing matrix, beta
}
