% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_EIR}
\alias{compute_EIR}
\title{Compute the EIR}
\usage{
compute_EIR(deout, pars)
}
\arguments{
\item{deout}{a \link{matrix}, the output of deSolve}

\item{pars}{a \link{list}}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve,
compute the EIR at every point in time
}
