% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_EIR_ty}
\alias{compute_EIR_ty}
\title{Compute the EIR}
\usage{
compute_EIR_ty(t, y, pars)
}
\arguments{
\item{t}{\link{numeric}}

\item{y}{\link{vector} the state variables}

\item{pars}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Using the time and a vector of variables,
compute the EIR
}
