% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_kappa}
\alias{compute_kappa}
\title{Compute kappa}
\usage{
compute_kappa(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This function computes kappa for each vector species in each patch
}
