% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_fraction.R
\name{compute_local_frac}
\alias{compute_local_frac}
\title{Compute the local fraction}
\usage{
compute_local_frac(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
