% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_terms_steady}
\alias{compute_terms_steady}
\title{Compute dynamical terms}
\usage{
compute_terms_steady(y_eq, pars)
}
\arguments{
\item{y_eq}{a \link{matrix}, the output of \code{rootSolve}}

\item{pars}{a \link{list}}
}
\value{
\link{matrix}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{root_solve}
}
