% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{dHdt.dynamic}
\alias{dHdt.dynamic}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{dHdt}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nStrata} or of length 0
}
\description{
Implements \link{dHdt} for the dynamic demography model
}
