% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-null.R
\name{dHdt.null}
\alias{dHdt.null}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{dHdt}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of 0s or of length 0
}
\description{
Implements \link{dHdt} for the null model.
}
