% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{dHdt}
\alias{dHdt}
\title{Derivatives of demographic changes in human populations}
\usage{
dHdt(pars, ...)
}
\arguments{
\item{pars}{an \link{environment}}

\item{...}{additional arguments which must be of the form \code{D}, \code{Y}, etc. \code{D}
is a \code{nStrata} by \code{nStrata} demography matrix and \code{Y} is a state vector. There
should be at least one pair of these passed to the function.}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of \code{pars$Hpar}.
}
