% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{dHdt}
\alias{dHdt}
\title{Derivatives of demographic changes in human populations}
\usage{
dHdt(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of \code{y}.
}
