% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{dHdt.static}
\alias{dHdt.static}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{dHdt}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length 0
}
\description{
Implements \link{dHdt} when \code{y} is static
}
