% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-trace.R
\name{dHdt.trace}
\alias{dHdt.trace}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{dHdt}{trace}(pars, ...)
}
\arguments{
\item{pars}{an \link{environment}}

\item{...}{additional arguments which must be of the form \code{D}, \code{Y}, etc. \code{D}
is a \code{nStrata} by \code{nStrata} demography matrix and \code{Y} is a state vector. There
should be at least one pair of these passed to the function.}
}
\value{
a \link{numeric} vector of length \code{nStrata} or of length 0
}
\description{
Implements \link{dHdt} for the forced (trace) model.
}
