% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{dLdt.basic}
\alias{dLdt.basic}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dLdt}{basic}(t, y, pars, eta)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}

\item{eta}{vector giving number of eggs being laid in each larval habitat}
}
\value{
a [numeric] vector
}
\description{
Implements [dLdt] for the basic competition model.
}
