% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{dLdt}
\alias{dLdt}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
dLdt(t, y, pars, eta)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{eta}{vector giving number of eggs being laid in each larval habitat}
}
\value{
a \link{numeric} vector of length \code{pars$L_ix}
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
