% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{dLdt.trace}
\alias{dLdt.trace}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dLdt}{trace}(t, y, pars, eta)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{eta}{vector giving number of eggs being laid in each larval habitat}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dLdt} for the trace (forced emergence) model.
}
