% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{dMYZdt.GeRM_dde}
\alias{dMYZdt.GeRM_dde}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYZdt}{GeRM_dde}(t, y, pars, Lambda, kappa)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the GeRM DDE model.
}
