% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{dMYZdt.Gtrace}
\alias{dMYZdt.Gtrace}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dMYZdt}{Gtrace}(t, y, pars, Lambda, kappa, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the Gtrace (forced emergence) model.
}
