% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dMYZdt}
\alias{dMYZdt}
\title{Derivatives for adult mosquitoes}
\usage{
dMYZdt(t, y, pars, Lambda, kappa, MosyBehavior)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}

\item{MosyBehavior}{values returned by \link{MosquitoBehavior}, potentially modified by control \link{VectorControl}}
}
\value{
a \link{numeric} vector
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
