% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{dMYZdt.RM_dde}
\alias{dMYZdt.RM_dde}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYZdt}{RM_dde}(t, y, pars, Lambda, kappa)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the RM DDE model.
}
