% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{dMYZdt.Ztrace}
\alias{dMYZdt.Ztrace}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dMYZdt}{Ztrace}(t, y, pars, Lambda, kappa)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{Lambda}{emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness of human population}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the Ztrace (forced emergence) model.
}
