% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{dXdt.hMoI}
\alias{dXdt.hMoI}
\title{Derivatives for human population}
\usage{
\method{dXdt}{hMoI}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the hybrid MoI model.
}
