% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{dXdt.hMoI}
\alias{dXdt.hMoI}
\title{Derivatives for human population}
\usage{
\method{dXdt}{hMoI}(t, y, pars, EIR)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{EIR}{vector giving the per-capita entomological inoculation rate for each strata}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the hybrid MoI model.
}
