% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{dXdt}
\alias{dXdt}
\title{Derivatives for human population}
\usage{
dXdt(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
