% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{dXdt.SIP}
\alias{dXdt.SIP}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SIP}(t, y, pars, FoI)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{FoI}{vector giving the per-capita force of infection for each strata}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SIP model.
}
