% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dXdt.SISdXdH}
\alias{dXdt.SISdXdH}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SISdXdH}(t, y, pars, EIR)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{EIR}{vector giving the per-capita entomological inoculation rate for each strata}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SIS model with demography.
}
