% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{EIP}
\alias{EIP}
\title{Compute the EIP}
\usage{
EIP(t, EIPmod)
}
\arguments{
\item{t}{current simulation time}

\item{EIPmod}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
This method dispatches on the type of \code{EIPmod}.
}
