% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous-null.R
\name{ExogenousForcing.null}
\alias{ExogenousForcing.null}
\title{Modify parameters due to exogenous forcing}
\usage{
\method{ExogenousForcing}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{ExogenousForcing} for the null model of exogenous forcing (do nothing)
}
