% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure}
\alias{Exposure}
\title{Exposure and Infection}
\usage{
Exposure(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object as a \link{list}}
}
\value{
the function modifies \strong{pars} and returns it: the computed FoI are stored as \code{pars$FoI}
}
\description{
This function translates the local daily entomological
inoculation rate (dEIR), computed as one of the transmission terms
into a measure of the daily force of infection (dFoI). The daily FoI
is the sum of two terms: 1) a function \link{F_foi} computes the local dFoI;
2) a function \link{travel_malaria} that computes the FoI resulting from
exposure while traveling.
}
