% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure}
\alias{Exposure}
\title{A model for exposure. The function \code{F_b} must be define}
\usage{
Exposure(t, y, pars, EIR, i)
}
\arguments{
\item{t}{the time}

\item{y}{the variables}

\item{pars}{a \link{list}}

\item{EIR}{is the daily eir}

\item{i}{the host species index}
}
\description{
A model for exposure. The function \code{F_b} must be define
}
