% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{F_alpha}
\alias{F_alpha}
\title{Number of newly emerging adults from each larval habitat}
\usage{
F_alpha(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
