% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{F_alpha.trace}
\alias{F_alpha.trace}
\title{Number of newly emerging adults from each larval habitat}
\usage{
\method{F_alpha}{trace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_alpha} for the trace (forced emergence) model.
}
