% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_beta}
\alias{F_beta}
\title{Compute the biting distribution matrix, beta}
\usage{
F_beta(t, y, pars, s, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars, }{a \link{list}}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}
}
