% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission-interface.R
\name{F_beta}
\alias{F_beta}
\title{Biting distribution matrix}
\usage{
F_beta(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars, }{a \link{list}}
}
\value{
pars, a \link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}
}
