% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodfeeding.R
\name{F_beta}
\alias{F_beta}
\title{Biting distribution matrix}
\usage{
F_beta(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$BFpar}
}
