% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_EIR.hMoI}
\alias{F_EIR.hMoI}
\title{Entomological inoculation rate on human strata}
\usage{
F_EIR.hMoI(t, y, pars, MosyBehavior)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{MosyBehavior}{values returned by \link{MosquitoBehavior}, potentially modified by control \link{VectorControl}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_EIR} for the hybrid MoI model.
}
