% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{F_EIR}
\alias{F_EIR}
\title{Entomological inoculation rate on human strata}
\usage{
F_EIR(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
