% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_EIR}
\alias{F_EIR}
\title{Entomological inoculation rate on human strata}
\usage{
F_EIR(t, y, pars, beta)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{beta, }{a \link{matrix}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the daily EIR for all the strata at time \code{t}
}
