% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{F_EIR.RM}
\alias{F_EIR.RM}
\title{Entomological inoculation rate on human strata}
\usage{
\method{F_EIR}{RM}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_EIR] for the generalized RM model.
}
