% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission-trace.R
\name{F_EIR.trace}
\alias{F_EIR.trace}
\title{Entomological inoculation rate on human strata}
\usage{
\method{F_EIR}{trace}(t, y, pars, MosyBehavior)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{MosyBehavior}{a \link{list} returned from \code{MosquitoBehavior}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_EIR} for when X is an trace model.
}
