% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{F_foi}
\alias{F_foi}
\title{A model for exposure}
\usage{
F_foi(eir, pars)
}
\arguments{
\item{eir}{the daily eir}

\item{pars}{a \link{list}}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of \code{pars$FOIpar}.
}
