% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_H.hMoI}
\alias{F_H.hMoI}
\title{Size of the human population}
\usage{
\method{F_H}{hMoI}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the hybrid MoI model.
}
