% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-trace.R
\name{F_H.trace}
\alias{F_H.trace}
\title{Size of human population denominators}
\usage{
\method{F_H}{trace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the forced (trace) model.
}
