% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_kappa}
\alias{F_kappa}
\title{Net infectiousness of human population to mosquitoes}
\usage{
F_kappa(t, y, pars, i, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the net infectiousness of humans in each patch at time \code{t}
}
