% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_kappa}
\alias{F_kappa}
\title{Net infectiousness of human population to mosquitoes}
\usage{
F_kappa(t, y, pars, beta)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{beta, }{a \link{matrix}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{pars$MYZpar$xde}.
}
