% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{F_kappa.RM_ode}
\alias{F_kappa.RM_ode}
\title{Net infectiousness of human population to mosquitoes}
\usage{
\method{F_kappa}{RM_ode}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_kappa} for the generalized RM ODE model.
}
