% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{F_nu.type2}
\alias{F_nu.type2}
\title{Type 2 functional response for the blood feeding rate}
\usage{
\method{F_nu}{type2}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_nu} for a static model
}
