% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-functions.R
\name{F_sugar.static}
\alias{F_sugar.static}
\title{Static model for sugar blood hosts}
\usage{
\method{F_sugar}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_sugar} for a static model
}
