% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_x.hMoI}
\alias{F_x.hMoI}
\title{Size of effective infectious human population}
\usage{
\method{F_x}{hMoI}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_x} for the hybrid MoI model.
}
