% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{F_x}
\alias{F_x}
\title{Size of effective infectious human population}
\usage{
F_x(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
