% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_x.SIP}
\alias{F_x.SIP}
\title{Size of effective infectious human population}
\usage{
\method{F_x}{SIP}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a [list]}
}
\value{
a [numeric] vector of length `nStrata`
}
\description{
Implements [F_x] for the SIP model.
}
