% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_X.SIS}
\alias{F_X.SIS}
\title{Size of effective infectious human population}
\usage{
\method{F_X}{SIS}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_X} for the SIS model.
}
