% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_x_tau.SIS}
\alias{F_x_tau.SIS}
\title{Size of lagged effective infectious human population}
\usage{
\method{F_x_tau}{SIS}(t, y, pars, tau)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{tau}{duration of lag \code{t-tau}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_x_tau} for the SIS model.
}
