% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inits}
\alias{get_inits}
\title{Get the initial values as a vector}
\usage{
get_inits(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the initial values as a vector
}
