% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{get_inits_H}
\alias{get_inits_H}
\title{Return initial values as a vector}
\usage{
get_inits_H(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Hpar}.
}
