% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{get_inits_L}
\alias{get_inits_L}
\title{Return initial values as a vector}
\usage{
get_inits_L(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
