% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{get_inits_MYZ.GeRM_dde}
\alias{get_inits_MYZ.GeRM_dde}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{GeRM_dde}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{get_inits_MYZ} for the GeRM model.
}
