% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{get_inits_X.SIP}
\alias{get_inits_X.SIP}
\title{Return initial values as a vector}
\usage{
\method{get_inits_X}{SIP}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
