% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{get_inits_X.SIS}
\alias{get_inits_X.SIS}
\title{Return initial values as a vector}
\usage{
\method{get_inits_X}{SIS}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
