% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{Hydrology}
\alias{Hydrology}
\title{Set the values of exogenous variables describing hydrology}
\usage{
Hydrology(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$HYDROLOGY}.
}
