% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{make_index_H.dynamic}
\alias{make_index_H.dynamic}
\title{Add indices for human population denominators to parameter list}
\usage{
\method{make_index_H}{dynamic}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_index_H} for dynamic models.
}
