% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basic.R
\name{make_index_L.basic}
\alias{make_index_L.basic}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_index_L}{basic}(pars)
}
\arguments{
\item{pars}{a [list]}
}
\value{
the modified parameter [list]
}
\description{
Implements [make_index_L] for basic competition model.
}
