% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{make_index_L.trace}
\alias{make_index_L.trace}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_index_L}{trace}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
the modified parameter \link{list}
}
\description{
Implements \link{make_index_L} for trace (forced emergence) model.
}
