% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{make_indices_MYZ}
\alias{make_indices_MYZ}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
make_indices_MYZ(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
