% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_indices_X.SIP}
\alias{make_indices_X.SIP}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{SIP}(pars)
}
\arguments{
\item{pars}{an \link{environment}}
}
\value{
none
}
\description{
Implements \link{make_indices_X} for the SIP model.
}
