% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_inits_X_SIS}
\alias{make_inits_X_SIS}
\title{Make inits for SIS human model}
\usage{
make_inits_X_SIS(pars, I0)
}
\arguments{
\item{pars}{a \link{list}}

\item{I0}{size of infected population in each strata}
}
\value{
none
}
\description{
Make inits for SIS human model
}
