% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_MYZpar_GeRM}
\alias{make_MYZpar_GeRM}
\title{Make parameters for a GeRM ODE adult mosquito model}
\usage{
make_MYZpar_GeRM(
  pars,
  MYZopts = list(),
  calK,
  solve_as = "dde",
  g = 1/12,
  setup_Fg = list(),
  sigma = 1/8,
  setup_Fsigma = list(),
  f = 0.3,
  setup_Ff = list(),
  q = 0.95,
  setup_Fq = list(),
  nu = 1,
  setup_Fnu = list(),
  eip = 11,
  setup_Feip = list(),
  eggsPerBatch = 60
)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZopts}{a \link{list} to overwrite defaults}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{solve_as}{is either \code{ode} to solve as an ode or \code{dde} to solve as a dde}

\item{g}{mosquito mortality rate}

\item{setup_Fg}{a \link{list} to set up F_g}

\item{sigma}{emigration rate}

\item{setup_Fsigma}{a \link{list} to set up F_sigma}

\item{f}{feeding rate}

\item{setup_Ff}{a \link{list} to set up F_f}

\item{q}{human blood fraction}

\item{setup_Fq}{a \link{list} to set up F_q}

\item{nu}{oviposition rate, per mosquito}

\item{setup_Fnu}{a \link{list} to set up F_nu}

\item{eip}{length of extrinsic incubation period}

\item{setup_Feip}{a \link{list} to set up F_eip}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
none
}
\description{
Make parameters for a GeRM ODE adult mosquito model
}
