% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloodfeeding.R
\name{make_parameters_BF_static}
\alias{make_parameters_BF_static}
\title{Sets up beta for a static matrix}
\usage{
make_parameters_BF_static(params)
}
\value{
a \link{matrix} of dimensions \code{nStrata} by \code{nPatches}
}
\description{
Implements \link{F_beta_lag} for dynamic denominators
}
