% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-basic.R
\name{make_parameters_demography_basic}
\alias{make_parameters_demography_basic}
\title{Make parameters for forced (basic) human demography model}
\usage{
make_parameters_demography_basic(
  pars,
  H,
  membershipH,
  searchWtsH,
  TimeSpent,
  F_birth,
  birthrate,
  deathrate,
  birthsXstrata
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{H}{a function taking a single argument \code{t} and returning a vector of length
\code{nStrata}.}

\item{membershipH}{is a vector describing patch residency}

\item{searchWtsH}{is a vector describing blood feeding search weights}

\item{TimeSpent}{is a matrix describing time spent among patches}

\item{F_birth}{is function that returns the birth rate as a function of time}

\item{birthrate}{is a parameter describing the current birthrate for each stratum}

\item{deathrate}{is a parameter describing the current deathrate for each stratum}

\item{birthsXstrata}{is a parameter describing which strata get births}
}
\value{
none
}
\description{
Make parameters for forced (basic) human demography model
}
