% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{make_parameters_demography_dynamic}
\alias{make_parameters_demography_dynamic}
\title{Make parameters for forced (trace) human demography model}
\usage{
make_parameters_demography_dynamic(pars, b, d, m, H0)
}
\arguments{
\item{pars}{an \link{environment}}

\item{b}{a list containing vectors of birth rates for each state in \eqn{\mathcal{X}}; each vector should be of length \code{nStrata}}

\item{d}{a list containing vectors of death rates for each state in \eqn{\mathcal{X}}; each vector should be of length \code{nStrata}}

\item{m}{a list containing vectors of ageing rates for each state in \eqn{\mathcal{X}}; each vector should be of length \code{nStrata-1}}

\item{H0}{initial size of human population in each strata at time 0}
}
\value{
none
}
\description{
Make parameters for forced (trace) human demography model
}
