% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-dynamic.R
\name{make_parameters_demography_dynamic}
\alias{make_parameters_demography_dynamic}
\title{Make parameters for forced (dynamic) human demography model}
\usage{
make_parameters_demography_dynamic(
  pars,
  H,
  membershipH,
  searchWtsH,
  TimeSpent,
  birthF,
  birthrate,
  Hmatrix,
  birthsXstrata
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{H}{a function taking a single argument \code{t} and returning a vector of length
\code{nStrata}.}

\item{membershipH}{is a vector describing patch residency}

\item{searchWtsH}{is a vector describing blood feeding search weights}

\item{TimeSpent}{is a matrix describing time spent among patches}

\item{birthF}{dispatches \code{F_birth}}

\item{birthrate}{is the population birthrate}

\item{Hmatrix}{does a set of state transitions}

\item{birthsXstrata}{distributes births to the youngest strata}
}
\value{
none
}
\description{
Make parameters for forced (dynamic) human demography model
}
