% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{make_parameters_demography_null}
\alias{make_parameters_demography_null}
\title{Make parameters for null human demography model}
\usage{
make_parameters_demography_null(pars, H, residence, searchWts, TaR)
}
\arguments{
\item{pars}{a \link{list}}

\item{H}{size of human population in each strata}

\item{residence}{is a vector describing patch residency}

\item{searchWts}{is a vector describing blood feeding search weights}

\item{TaR}{is a matrix describing time spent among patches}
}
\value{
none
}
\description{
Make parameters for null human demography model
}
