% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-interface.R
\name{make_parameters_demography_null}
\alias{make_parameters_demography_null}
\title{Make parameters for null human demography model}
\usage{
make_parameters_demography_null(pars, H)
}
\arguments{
\item{pars}{a \link{list}}

\item{H}{size of human population in each strata}
}
\value{
none
}
\description{
Make parameters for null human demography model
}
