% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography-static.R
\name{make_parameters_demography_static}
\alias{make_parameters_demography_static}
\title{Make parameters for static human demography model}
\usage{
make_parameters_demography_static(
  pars,
  H,
  residence,
  searchWts,
  TaR,
  birthFpars,
  Hmatrix,
  birthsXstrata
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{H}{size of human population in each strata}

\item{residence}{is a vector describing patch residency}

\item{searchWts}{is a vector describing blood feeding search weights}

\item{TaR}{is a matrix describing time spent among patches}

\item{birthFpars}{setup to dispatch and compute \code{F_birth}}

\item{Hmatrix}{does a set of state transitions}

\item{birthsXstrata}{distributes births to the youngest strata}
}
\value{
none
}
\description{
Make parameters for static human demography model
}
