% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{make_parameters_exposure_nb}
\alias{make_parameters_exposure_nb}
\title{Make parameters for the null model of exposure}
\usage{
make_parameters_exposure_nb(pars, sz)
}
\arguments{
\item{pars}{a \link{list}}

\item{sz}{the size parameter, as in dnbinom(mu=mu, size=size)}
}
\value{
none
}
\description{
Make parameters for the null model of exposure
}
