% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{make_parameters_L_trace}
\alias{make_parameters_L_trace}
\title{Make parameters for trace aquatic mosquito model}
\usage{
make_parameters_L_trace(Lambda)
}
\arguments{
\item{Lambda}{vector of emergence rates from each aquatic habitat}
}
\value{
a \link{list} with class \code{trace}.
}
\description{
Make parameters for trace aquatic mosquito model
}
