% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{make_parameters_L_trace}
\alias{make_parameters_L_trace}
\title{Make parameters for trace aquatic mosquito model}
\usage{
make_parameters_L_trace(
  pars,
  Lambda,
  Lt = function(t, pars) {
     1
 }
)
}
\arguments{
\item{pars}{an \link{environment}}

\item{Lambda}{vector of mean emergence rates from each aquatic habitat}

\item{Lt}{is a \link{function} of the form Lt(t,pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for trace aquatic mosquito model
}
