% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_parameters_MYZ_GeRM}
\alias{make_parameters_MYZ_GeRM}
\title{Make parameters for GeRM ODE adult mosquito model}
\usage{
make_parameters_MYZ_GeRM(pars, calK, solve_as = "dde")
}
\arguments{
\item{pars}{an \link{environment}}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{solve_as}{is either \code{ode} to solve as an ode or \code{dde} to solve as a dde}
}
\value{
none
}
\description{
Make parameters for GeRM ODE adult mosquito model
}
