% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{make_parameters_MYZ_Ztrace}
\alias{make_parameters_MYZ_Ztrace}
\title{Make parameters for Ztrace aquatic mosquito model}
\usage{
make_parameters_MYZ_Ztrace(pars, Zo = 1, Zf = NULL)
}
\arguments{
\item{pars}{an \link{environment}}

\item{Zo}{a vector of mosquito densities}

\item{Zf}{a \link{function} that computes Z(t, pars)}
}
\value{
none
}
\description{
Make parameters for Ztrace aquatic mosquito model
}
