% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{make_parameters_MYZ_Ztrace}
\alias{make_parameters_MYZ_Ztrace}
\title{Make parameters for Ztrace aquatic mosquito model}
\usage{
make_parameters_MYZ_Ztrace(pars, Zm, f, q, Zf = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{Zm}{a vector of mean mosquito densities}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Zf}{a \link{function} of the form Zf(t, pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for Ztrace aquatic mosquito model
}
