% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_parameters_X_SIS}
\alias{make_parameters_X_SIS}
\title{Make parameters for SIS human model}
\usage{
make_parameters_X_SIS(b, c, r, X0, H)
}
\arguments{
\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}

\item{X0}{size of infected population in each strata}

\item{H}{size of human population in each strata}
}
\value{
a \link{list} with class \code{SIS}.
}
\description{
Make parameters for SIS human model
}
