% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_parameters_X_SIS}
\alias{make_parameters_X_SIS}
\title{Make parameters for SIS human model}
\usage{
make_parameters_X_SIS(pars, b, c, r, Psi, wf = 1, X0, H)
}
\arguments{
\item{pars}{an \link{environment}}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}

\item{Psi}{a \link{matrix} of dimensions \code{nPatches} by \code{nStrata}}

\item{wf}{vector of biting weights of length \code{nStrata}}

\item{X0}{size of infected population in each strata}

\item{H}{size of human population in each strata}
}
\value{
none
}
\description{
Make parameters for SIS human model
}
