% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_parameters_xde}
\alias{make_parameters_xde}
\title{Make base parameters}
\usage{
make_parameters_xde(solve_as = "ode")
}
\arguments{
\item{solve_as, }{either "ode" or "dde"}
}
\value{
a \link{list}
}
\description{
Make base parameters
}
