% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeatrisk.R
\name{make_TaR}
\alias{make_TaR}
\title{Make a mosquito dispersal matrix, called TaR}
\usage{
make_TaR(nPatches, residence, TaR, opts = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the residence vector}

\item{TaR}{a matrix; or a string}

\item{opts}{a list of options to configure TaR}
}
\value{
a \link{matrix}
}
\description{
Make a mosquito dispersal matrix, called TaR
}
