% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_Xinits_SIP}
\alias{make_Xinits_SIP}
\title{Make initial values for the SIP human model, with defaults}
\usage{
make_Xinits_SIP(pars, Xopts = list(), S0 = NULL, I0 = 1, P0 = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{S0}{the initial values of the parameter S}

\item{I0}{the initial values of the parameter I}

\item{P0}{the initial values of the parameter P}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIP human model, with defaults
}
