% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_Xpar_SIS}
\alias{make_Xpar_SIS}
\title{Make parameters for SIS human model, with defaults}
\usage{
make_Xpar_SIS(nStrata, Xopts = list(), b = 0.55, r = 1/180, c = 0.15)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{c}{transmission probability (efficiency) from human to mosquito}
}
\value{
a \link{list}
}
\description{
Make parameters for SIS human model, with defaults
}
