% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation-interface.R
\name{MalariaImportation}
\alias{MalariaImportation}
\title{Malaria importation through infected visitors, etc.}
\usage{
MalariaImportation(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
pars a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MIpar}.
}
