% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{MosquitoBehavior.Ztrace}
\alias{MosquitoBehavior.Ztrace}
\title{Compute bloodfeeding and mortality rates}
\usage{
\method{MosquitoBehavior}{Ztrace}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a named \link{list}
}
\description{
Implements \link{MosquitoBehavior} for the Ztrace model.
}
