% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{ResourceAvailability.GeRM}
\alias{ResourceAvailability.GeRM}
\title{Set the availability of resources}
\usage{
\method{ResourceAvailability}{GeRM}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{ResourceAvailability} for the GeRM model
}
