% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-null.R
\name{ResourceAvailability.null}
\alias{ResourceAvailability.null}
\title{Modify parameters due to exogenous forcing}
\usage{
\method{ResourceAvailability}{null}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Implements \link{ResourceAvailability} for the null model of hydrology (do nothing)
}
