% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_eip_static}
\alias{setup_eip_static}
\title{Set up a static model for the EIP}
\usage{
setup_eip_static(EIPopts = list(), eip = 11)
}
\arguments{
\item{EIPopts}{a \link{list}}

\item{eip}{the extrinsic incubation period (in days)}
}
\value{
\link{list}
}
\description{
Set up a static model for the EIP
}
