% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-static.R
\name{setup_resources_static}
\alias{setup_resources_static}
\title{Set up parameters for a static model for resource availability}
\usage{
setup_resources_static(pars, RAopts, other = 0, sugar = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{RAopts}{a set of options that would override the defaults}

\item{other}{the availability of other blood hosts}

\item{sugar}{the availability of sugar}
}
\value{
none
}
\description{
Set up parameters for a static model for resource availability
}
