% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorcontrol-null.R
\name{setup_vc_null}
\alias{setup_vc_null}
\title{Set up the vector control null model (do nothing)}
\usage{
setup_vc_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Set up the vector control null model (do nothing)
}
