% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-null.R
\name{setup_visitors_null}
\alias{setup_visitors_null}
\title{Make parameters for the null model visitors (no visitors)}
\usage{
setup_visitors_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model visitors (no visitors)
}
