% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_weather_null}
\alias{setup_weather_null}
\title{Set up the null model for weather}
\usage{
setup_weather_null(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the null model for weather
}
