% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{setup_Xpar}
\alias{setup_Xpar}
\title{A function to set up Xpar}
\usage{
setup_Xpar(pars, Xname, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{Xname}{a \link{character} string}

\item{Xopts}{a \link{list}}
}
\value{
none
}
\description{
This method dispatches on \code{Xname}.
}
