% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{setup_Xpar.trace}
\alias{setup_Xpar.trace}
\title{Setup Xpar.trace}
\usage{
\method{setup_Xpar}{trace}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the trace model
}
