% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits-interface.R
\name{SugarBaitEffectSizes}
\alias{SugarBaitEffectSizes}
\title{Methods for the effect sizes of the sugar baits}
\usage{
SugarBaitEffectSizes(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$SUGAR_BAITS}.
}
