% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_EggLaying.dynamic}
\alias{update_EggLaying.dynamic}
\title{Update the total availability of aquatic habitats and egg laying}
\usage{
\method{update_EggLaying}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{an \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{update_EggLaying} for a dynamic model
}
