% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{update_inits_L}
\alias{update_inits_L}
\title{Set the initial values from a vector of model states}
\usage{
update_inits_L(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
