% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{update_inits_MYZ.GeRM_dde}
\alias{update_inits_MYZ.GeRM_dde}
\title{Make inits for RM adult mosquito model}
\usage{
\method{update_inits_MYZ}{GeRM_dde}(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
Make inits for RM adult mosquito model
}
