% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{update_inits_MYZ}
\alias{update_inits_MYZ}
\title{Set the initial values as a vector}
\usage{
update_inits_MYZ(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
