% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{update_inits_MYZ.RM_dde}
\alias{update_inits_MYZ.RM_dde}
\title{Update inits for RM adult mosquito model}
\usage{
\method{update_inits_MYZ}{RM_dde}(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
Update inits for RM adult mosquito model
}
