% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{update_inits_X.SIS}
\alias{update_inits_X.SIS}
\title{Update inits for the SIS human model from a vector of states}
\usage{
\method{update_inits_X}{SIS}(pars, y0)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}
}
\value{
none
}
\description{
Update inits for the SIS human model from a vector of states
}
