% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_O.dynamic}
\alias{update_O.dynamic}
\title{Update other host availability dynamically}
\usage{
update_O.dynamic(t, pars)
}
\value{
a \link{list}
}
\description{
Implements \link{update_O} for a static model
}
