% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources-dynamic.R
\name{update_W.dynamic}
\alias{update_W.dynamic}
\title{Update host availability dynamically}
\usage{
update_W.dynamic(t, y, pars)
}
\value{
a \link{list}
}
\description{
Implements \link{update_W} for a static model
}
