% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-basic.R
\name{Visitors.basic}
\alias{Visitors.basic}
\title{Visitors, the basic model}
\usage{
\method{Visitors}{basic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{Visitors} for the basic model for Visitors
}
