% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.null}
\alias{Weather.null}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{null}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Weather} for the null model (no variables)
}
