% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqn.R
\name{xDE_diffeqn}
\alias{xDE_diffeqn}
\title{Generalized spatial differential equation model}
\usage{
xDE_diffeqn(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
