% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffeqn.R
\name{xDE_diffeqn_mosy}
\alias{xDE_diffeqn_mosy}
\title{Generalized spatial differential equation model (mosquito only)}
\usage{
xDE_diffeqn_mosy(t, y, pars, kappa, MosyBehavior)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \link{environment}}

\item{kappa}{a vector}

\item{MosyBehavior}{a \link{list} emulating the output of \link{MosquitoBehavior} for
the appropriate adult mosquito model}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Mirrors \link{xDE_diffeqn} but only includes the adult and aquatic
mosquito components.
}
