% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{xde_lines_PR}
\alias{xde_lines_PR}
\title{Add lines for the prevalence / parasite rate (PR) from a model of human infection and immunity}
\usage{
xde_lines_PR(time, PR, nStrata, clrs = "black", llty = 1)
}
\arguments{
\item{time}{the time}

\item{PR}{the computed parasite rate}

\item{nStrata}{the number of human / host population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) that specifies \code{lty} for plotting}
}
\description{
Add lines for the prevalence / parasite rate (PR) from a model of human infection and immunity
}
