% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-MYZ.R
\name{xde_plot_YZ}
\alias{xde_plot_YZ}
\title{Plot the density of infected and infective mosquitoes}
\usage{
xde_plot_YZ(
  pars,
  Yclrs = "purple",
  Zclrs = "darkred",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a \link{list} that defines an \code{exDE} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{Yclrs}{a \link{character} vector of colors for infected mosquitoes}

\item{Zclrs}{a \link{character} vector of colors for infective mosquitoes}

\item{llty}{an \link{integer} (or integers) to set the \code{lty} for plotting}

\item{stable}{a \link{logical} set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a \link{logical} to plot axes only if TRUE}
}
\description{
Plot the density of infected and infective mosquitoes
}
