% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xde_setup_cohort}
\alias{xde_setup_cohort}
\title{Set up a model for xde_diffeqn_cohort}
\usage{
xde_setup_cohort(
  F_eir,
  modelName = "unnamed",
  Xname = "SIS",
  HPop = 1000,
  searchB = 1,
  Xopts = list(),
  Hopts = list()
)
}
\arguments{
\item{F_eir}{is a function F_eir(t, pars) that returns the daily FoI}

\item{modelName}{is a name for the model (arbitrary)}

\item{Xname}{is a character string defining a X model}

\item{HPop}{is the number of humans in each patch}

\item{searchB}{is a vector of search weights for blood feeding}

\item{Xopts}{a list to configure the X model}

\item{Hopts}{a list to configure the H model}
}
\value{
a \link{list}
}
\description{
Set up a model for xde_diffeqn_cohort
}
