% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xde_setup_human}
\alias{xde_setup_human}
\title{Set up a model for xde_diffeqn_human}
\usage{
xde_setup_human(
  modelName,
  Xname = "SIS",
  HPop = 1000,
  MYZopts = list(),
  Xopts = list(),
  Hopts = list(),
  residence = 1,
  searchB = 1,
  TaR = "athome",
  TaRopts = list(),
  eir_out = TRUE,
  fqZ_out = FALSE,
  NI_out = FALSE,
  kappa_out = TRUE
)
}
\arguments{
\item{modelName}{is a name for the model (arbitrary)}

\item{Xname}{is a character string defining a X model}

\item{HPop}{is the number of humans in each patch}

\item{MYZopts}{a list to configure the MYZ model}

\item{Xopts}{a list to configure the X model}

\item{Hopts}{a list to configure the H model}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{searchB}{is a vector of search weights for blood feeding}

\item{TaR}{is either a TaR matrix or a string to call a function that sets it up}

\item{TaRopts}{are the options to setup TaR}

\item{eir_out}{the cumulative eir is output if TRUE}

\item{fqZ_out}{the cumulative fqZ is output if TRUE}

\item{NI_out}{the cumulative NI is output if TRUE}

\item{kappa_out}{the cumulative kappa is output if TRUE}
}
\value{
a \link{list}
}
\description{
Set up a model for xde_diffeqn_human
}
